% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_operator.R
\name{BinaryOperator}
\alias{BinaryOperator}
\title{Binary Operator}
\usage{
BinaryOperator(operator)
}
\arguments{
\item{operator}{A binary operator. r binop_validity_message}
}
\value{
A binary operator consistent with the \code{C++} engine.
}
\description{
Convert a function that represents an elementwise binary
operator into one that is consistent with the \code{C++}
engine.
}
\examples{
set.seed(1L)
A = matrix(abs(rnorm(6)), 3, 2)  # 3 by 2 matrix
x = matrix(abs(rnorm(3)))        # 3 by 1 matrix
y = t(abs(rnorm(2)))             # 1 by 2 matrix
times = BinaryOperator(`*`)
pow = BinaryOperator(`^`)
identical(times(A, x), times(x, A))  ## TRUE
identical(pow(A, x), pow(x, A))  ## FALSE

}
