% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/products.R
\name{cartesian}
\alias{cartesian}
\title{Cartesian Product}
\usage{
cartesian(x, y)
}
\arguments{
\item{x}{Object with a \verb{$frame} method that returns a data frame.}

\item{y}{Object with a \verb{$frame} method that returns a data frame.}
}
\description{
Cartesian Product
}
