% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/products.R
\name{cartesian_self}
\alias{cartesian_self}
\title{Self-Cartesian Product}
\usage{
cartesian_self(x, left_filter, right_filter, .wrt, .rm = .wrt)
}
\arguments{
\item{x}{A \code{\link{Partition}} object.}

\item{left_filter}{Character string giving the name of the left filter.}

\item{right_filter}{Character string giving the name of the right filter.}

\item{.wrt}{Character string giving the name of a column in \code{x}.}

\item{.rm}{Character string giving the dot-concatenated names of columns
to remove from \code{x}.}
}
\description{
Self-Cartesian Product
}
