% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{Collection}
\alias{Collection}
\title{Collection}
\usage{
Collection(...)
}
\arguments{
\item{...}{Named components.}
}
\value{
Object of class \code{Collection} with the following methods.
\subsection{Methods}{
\itemize{
\item \verb{$get(component_name)}: Return components by name.
\item \verb{$freeze()}: Return the object itself so that the \verb{$freeze()} method
in \code{\link{Files}} objects feel idempotent.
}
}
}
\description{
Named collection of components.
}
\details{
Much like a \code{\link{list}} but designed to work with \code{\link{Files}}.
Typically one produces a \code{Collection} object by freezing a
\code{\link{Files}} object so that changes to the files that are
represented by the object will not affect the object.
}
