% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_and_labels.R
\name{comparison}
\alias{comparison}
\alias{all_equals}
\alias{all_consistent}
\title{Comparison Functions}
\usage{
all_equals(x, y)

all_consistent(x, y)
}
\arguments{
\item{x}{\code{\link{character}} object}

\item{y}{\code{\link{character}} object}
}
\description{
Comparison Functions
}
\section{Functions}{
\itemize{
\item \code{all_equals()}: Is it true that all corresponding elements of \code{x}
and \code{y} are equal, have the same shape, and have no missing values?

\item \code{all_consistent()}: Is it true that all corresponding elements of \code{x}
and \code{y} are either equal or at least one is a blank string, have the
same shape, and have no missing values?

}}
