% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_model.R
\name{ConstIntVecs}
\alias{ConstIntVecs}
\title{Constant Integer Vectors}
\usage{
ConstIntVecs(...)
}
\arguments{
\item{...}{Named arguments, each of which can be coerced to an integer vector.}
}
\description{
Make a list of integer vectors available for engine methods.
}
