% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data_structure.R
\name{DerivationExpander}
\alias{DerivationExpander}
\title{DerivationExpander}
\usage{
DerivationExpander(model)
}
\arguments{
\item{model}{Object of type \code{\link{Model}}
\subsection{Methods}{
\itemize{
\item \verb{$expand_derivation(derivation)} -- Expand a single derivation by name.
\item \verb{$expand_derivations()} -- Example all derivations in the model.
}
}}
}
\description{
Construct an object for expanding the derivations within a
\code{\link{Model}}.
}
