% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{DistrSpec}
\alias{DistrSpec}
\title{Distribution Specification}
\usage{
DistrSpec(distr_param_objs = list(), default_trans = list())
}
\arguments{
\item{distr_param_objs}{List of DistrParam objects defining the distribution. All
distributions must have a \code{location} parameter as the first parameter when
the distribution is instantiated. However, this \code{location} parameter
might be removed by the \code{remove_location_parameter} method, which
should be called later if it is determined that the distribution does
not require a \code{location} parameter (e.g., it is a likelihood component
that takes the simulated trajectory as the location).}

\item{default_trans}{list of default parameter transformation objects
of class DistrParamTrans for each \code{distr_param_objs}.}
}
\description{
Class representing the specification of a distribution, either a prior or
likelihood component.
}
\details{
Extend this class to develop new distributions.
}
