% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\docType{data}
\name{.known_dist}
\alias{.known_dist}
\title{Pick out obs/location pairs from terms involving probability distributions
find_obs_pairs(~ dnorm(a, b, c) + dpois(d, e))
find_obs_pairs(stuff ~ other_stuff + more_stuff)}
\format{
An object of class \code{character} of length 5.
}
\usage{
.known_dist
}
\description{
Pick out obs/location pairs from terms involving probability distributions
find_obs_pairs(~ dnorm(a, b, c) + dpois(d, e))
find_obs_pairs(stuff ~ other_stuff + more_stuff)
}
\keyword{datasets}
