% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_expr.R
\docType{data}
\name{empty_matrix}
\alias{empty_matrix}
\title{Empty Matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 0 rows and 0 columns.
}
\usage{
empty_matrix
}
\description{
Empty Matrix
}
\keyword{datasets}
