% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_expr.R
\docType{data}
\name{empty_trajectory}
\alias{empty_trajectory}
\title{Empty Trajectory}
\format{
A data frame with zero rows and the following columns: \code{matrix},
\code{time}, \code{row}, \code{col}, \code{value}.
}
\usage{
empty_trajectory
}
\description{
Output of \code{\link{mp_trajectory}} if nothing is simulated.
}
\keyword{datasets}
