% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_eval.R
\name{engine_eval}
\alias{engine_eval}
\title{Engine Evaluation}
\usage{
engine_eval(e, ..., .matrix_to_return)
}
\arguments{
\item{e}{Expression as a two-sided formula.}

\item{...}{Named objects that can be coerced into numeric matrices.}

\item{.matrix_to_return}{Optional name of one of the matrices given
in \code{...} to be returned. If this argument is missing, the
matrix that will be returned is the matrix on the left-hand-side of the
expression.}
}
\value{
Matrix being produced on the left-hand-side or matrix given in
\code{.matrix_to_return} if it is provided.
}
\description{
Evaluate an expression in the TMB-based C++ simulation and objective
function engine.
}
\details{
This function is useful for trying out functions that can be used to
define \code{macpan2} models.
}
\examples{
engine_eval(x ~ exp(y), y = pi) # ~ 23.14069

# It is not currently possible to assign values to a subset of
# a matrix in a natural way (e.g. you cannot do things like x[1] = exp(y)),
# but you can achieve this functionality using the assign function.
engine_eval(dummy ~ assign(x, 1, 0, exp(y))
  , dummy = empty_matrix
  , x = rep(0, 2)
  , y = pi
  , .matrix_to_return = "x"
)
}
