% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_methods.R
\name{EngineMethods}
\alias{EngineMethods}
\title{Engine Methods}
\usage{
EngineMethods(exprs = list(), int_vecs = IntVecs())
}
\arguments{
\item{exprs}{A named \code{list} of formulas that can be matched to one of the
method prototypes (TODO: describe these).}

\item{int_vecs}{An \code{\link{IntVecs}} object containing integer vectors
that can be used by the methods.}
}
\description{
List of methods for pre-processing matrices before returning or
assigning them. One benefit of these methods is that they can make use
of user-supplied integer vectors that are stored in C++ as integers. Another
benefit can be readability but this is subjective.
}
