% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivation_expander.R
\name{ExpandDerivationDefinitions}
\alias{ExpandDerivationDefinitions}
\title{Expand Derivation Definitions}
\usage{
ExpandDerivationDefinitions(derivation, variables, required_partitions)
}
\arguments{
\item{derivation}{Model derivations.}

\item{variables}{Model variables.}

\item{required_partitions}{Undotted names of partitions used when naming
variables.}
}
\description{
Expand Derivation Definitions
}
