% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_expr.R
\name{ExpressionFormula}
\alias{ExpressionFormula}
\title{Expression Formula}
\usage{
ExpressionFormula(expression_string, valid_vars)
}
\arguments{
\item{expression_string}{String giving a model expression}

\item{valid_vars}{Named list of variables that could be used in the
expression.}
}
\description{
Expression Formula
}
