% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{Files}
\alias{Files}
\title{Files}
\usage{
Files(directory, ...)
}
\arguments{
\item{directory}{Path to a directory containing the files.}

\item{...}{One or more \code{\link{reader_spec}}(s) that point to files
within the \code{directory} and define what \code{\link{Reader}} should
be used to access the file.}
}
\value{
Object of class \code{Collection} with the following methods.
\subsection{Methods}{
\itemize{
\item \verb{$get(component_name)}: Return components by name.
\item \verb{$freeze()}: Convert to a \code{\link{Collection}} object, which is
equivalent to a \code{Files} object in that the \code{$get} method returns
the same objects. The difference is that a \code{Files} object
will update what it returns if the associated files change on disk, whereas
\code{Collection} is intended to be (by convention) immutable (hence the
verb 'freeze'). The 'by convention' bit means that a user is free to change
the contents of the private \code{.components} field, but this would violate
the convention.
}
}
}
\description{
Construct objects for representing files on disk.
}
\details{
Files can be accessed by reading them into R objects in memory on demand,
and caching the objects. The cache is used to return file contents unless
the file is updated, in which case the file is accessed again and
corresponding object re-cached.
}
