% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_expr.R
\name{finalizer}
\alias{finalizer}
\alias{finalizer_char}
\alias{finalizer_index}
\title{Finalizers}
\usage{
finalizer_char(x)

finalizer_index(x)
}
\arguments{
\item{x}{Raw parsed expression.}
}
\description{
Finalizers
}
\section{Functions}{
\itemize{
\item \code{finalizer_char()}: Finalize parsed expression so that the parse table is
a little more human readable.

\item \code{finalizer_index()}: Finalize parsed expression so that the parse table can
be passed to the C++ engine.

}}
