% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_frame.R
\name{find_all_paths}
\alias{find_all_paths}
\title{Find all Paths}
\usage{
find_all_paths(edges_df, start_node_guesses = character(0L))
}
\arguments{
\item{edges_df}{A data frame with a \code{from} and a \code{to} column.}

\item{start_node_guesses}{Optional guesses for nodes that start
paths. This is useful for models that are not directed acyclic
graphs (DAGs).}
}
\value{
List of character vectors of state variable names, each
vector describing a path through the model.
}
\description{
Find all paths through a compartmental model.
}
