% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{fit_distr_params}
\alias{fit_distr_params}
\alias{mp_fit}
\alias{mp_nofit}
\title{Fitting Distributional Parameters}
\usage{
mp_fit(x, trans = DistrParamTransDefault())

mp_nofit(x, trans = DistrParamTransDefault())
}
\arguments{
\item{x}{numeric starting value of the distributional parameter to fit, or
character name of an existing variable in the model with a default starting
value to use.}

\item{trans}{transformation to apply to the distributional parameter.
By default, distributional parameters inherit a default transformation from
the associated distribution. For example, the standard deviation parameter
\code{sd} in the \code{\link{mp_normal}} distributions has a default log
transformation specified using \code{\link{mp_log}}.}
}
\description{
Distributional parameters can be added to the list of parameters that are fit
during calibration. By default, distributional parameters in priors and
likelihoods are not fit. Use \code{mp_nofit} to exclude distributional parameters
from being fit and \code{mp_fit} to fit them.
}
