% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_expander.R
\name{FlowExpander}
\alias{FlowExpander}
\title{Flow Expander}
\usage{
FlowExpander(model)
}
\arguments{
\item{model}{Object constructed with \code{\link{ModelFiles}} or
\code{\link{ModelCollection}}.}
}
\description{
Flow Expander
}
