% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_expr.R
\name{FormulaEnvironment}
\alias{FormulaEnvironment}
\title{Formula Environment}
\usage{
FormulaEnvironment(valid_vars)
}
\arguments{
\item{valid_vars}{Named list of variables that could be used in a
formula.}
}
\description{
Formula Environment
}
