% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{groupSums}
\alias{groupSums}
\title{Group Sums}
\usage{
groupSums(x, f, n)
}
\arguments{
\item{x}{A numeric column vector.}

\item{f}{A column vector the same length as \code{x}.
containing integers between \code{0} and \code{n-1}.}

\item{n}{Length of the output column vector.}
}
\description{
Column vector containing the
sums of groups of elements in \code{x}. The groups are
determined by the integers in \code{f} and the order of
the sums in the output is determined by these
integers.
}
