% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{Index}
\alias{Index}
\alias{Index.Partition}
\alias{Index.data.frame}
\alias{Index.Index}
\alias{print.Index}
\alias{names.Index}
\alias{labelling_names.Index}
\alias{to_labels.Index}
\alias{labels.Index}
\title{Index}
\usage{
Index(partition, labelling_names = names(partition), reference_index = NULL)

\method{Index}{Partition}(partition, labelling_names = names(partition), reference_index = NULL)

\method{Index}{data.frame}(partition, labelling_names = names(partition), reference_index = NULL)

\method{Index}{Index}(partition, labelling_names = names(partition), reference_index = NULL)

\method{print}{Index}(x, ...)

\method{names}{Index}(x)

\method{labelling_names}{Index}(x)

\method{to_labels}{Index}(x)

\method{labels}{Index}(x, ...)
}
\arguments{
\item{partition}{A data frame (or data frame-like) object containing
definitions of the index. This object can be a \code{\link{data.frame}},
\code{\link{Partition}}, or another \code{\link{Index}} object.}

\item{labelling_names}{A \code{\link{character}} vector of the names of the
index that will be used to label the model components (i.e. rows) being
described. The \code{labelling_names} cannot have duplicates and must
contain at least one name. The index given by the \code{labelling_names}
must uniquely identify each row.}

\item{reference_index}{(Advanced) An optional partition to use when
computing subset indices.}

\item{x}{\code{Index} object.}

\item{...}{For consistency with existing S3 methods.}
}
\description{
Make an object to represent descriptions of model components and their
dimensions of variation. \code{Index} objects generalize and wrap
\code{\link{data.frame}}s. Each row of this \code{\link{data.frame}}-like
object is an entry in the index, and each column provides a description of
the entries. For example, the following \code{Index} object describes the
state variables of an age-structured SIR model. Each row corresponds to a
state variable and each state variable is described by columns \code{Epi} and
\code{Age}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{#>  Epi   Age
#>    S young
#>    I young
#>    R young
#>    S   old
#>    I   old
#>    R   old
}\if{html}{\out{</div>}}
}
\section{Methods (by class)}{
\itemize{
\item \code{Index(Partition)}: Create a \code{Index} object from a
\code{\link{Partition}} object. These \code{\link{Partition}} objects
wrap \code{\link{data.frame}}s. These data frames must follow certain
restrictions.

\item \code{Index(data.frame)}: Create a \code{Index} object from a
\code{\link{data.frame}} object that follows certain restrictions
documented in the help page for \code{\link{Partition}}s.

\item \code{Index(Index)}: Create another \code{Index} object from an
existing \code{Index} object, possibly with new \code{labelling_names}.

}}
\section{Methods (by generic)}{
\itemize{
\item \code{print(Index)}: Print a \code{Index} object.

\item \code{names(Index)}: Get the names of the index
(i.e. columns) of a \code{Index} object.

\item \code{labelling_names(Index)}: Retrieve the \code{labelling_names} of
a \code{Index} object. These \code{labelling_names}
are the names of the index that are used to label the model
components.

\item \code{to_labels(Index)}: Convert a \code{Index} object into
a character vector giving labels associated with each model component
(i.e. row) being described.

\item \code{labels(Index)}: Convert a \code{Index} object into
a character vector giving labels associated with each model component
(i.e. row) being described.

}}
\examples{
sir = data.frame(
  Epi   = c("S",     "I",     "R",     "D"   ),
  Vital = c("alive", "alive", "alive", "dead")
) |> Index(labelling_names = "Epi")
age = data.frame(
  Age = c("young", "old")
) |> Index()
state_index = mp_cartesian(sir, age)
labels(state_index)
(state_index
  |> mp_subset(Vital = "alive")
  |> labels()
)

}
\seealso{
\code{\link[=mp_index]{mp_index()}}
}
