% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_data.R
\name{IndexedExpressions}
\alias{IndexedExpressions}
\title{Indexed Expressions}
\usage{
IndexedExpressions(
  ...,
  index_data,
  vector_list = list(),
  unstructured_matrix_list = list()
)
}
\arguments{
\item{...}{Formula objects that reference the columns in the
\code{index_data}, the vectors in \code{vector_list} and the matrices
in \code{unstructured_matrix_list}.}

\item{index_data}{An object produced using \code{\link{mp_link_data}}.}

\item{vector_list}{Named list of objected produced using
\code{\link{mp_vector}}.}

\item{unstructured_matrix_list}{Named list of objects that can be coerced
to a matrix.}
}
\description{
Indexed Expressions
}
