% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_expr.R
\name{initial_valid_vars}
\alias{initial_valid_vars}
\title{Initial Valid Variables}
\usage{
initial_valid_vars(valid_var_names)
}
\arguments{
\item{valid_var_names}{Character vector of variable names.}
}
\description{
Initial Valid Variables
}
