% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_methods.R
\name{integration_methods}
\alias{integration_methods}
\alias{Euler}
\alias{RK4}
\title{Integration Methods}
\usage{
Euler(model_simulator)

RK4(model_simulator)
}
\arguments{
\item{model_simulator}{Object of class \code{\link{TMBSimulator}}.}
}
\description{
Functions to pass as \code{integration_method}s to \code{\link{SimulatorConstructor}}.
}
\section{Functions}{
\itemize{
\item \code{Euler()}: Euler method.

\item \code{RK4()}: Runge Kutta 4 method.

}}
