% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const_int_vec.R
\name{IntVecs}
\alias{IntVecs}
\title{Constant Integer Vectors}
\usage{
IntVecs(...)
}
\arguments{
\item{...}{Named arguments, each of which can be coerced to an integer vector.}
}
\description{
Make a list of integer vectors available for engine methods.
}
