% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_products.R
\name{LabelMultiply}
\alias{LabelMultiply}
\title{Label Multiplication}
\usage{
LabelMultiply(times = "_", add = ", ")
}
\arguments{
\item{times}{String used to separate two strings being pasted together.}

\item{add}{String used to separate the items in a character vector being
collapsed into a character string.}
}
\value{
Object of class \code{LabelMultiply}.
\subsection{Methods}{
\itemize{
\item \verb{$outer(x, y)}: Outer product by pasting together every pairwise
combination of the elements in \code{x} with those in \code{y}.
\item \verb{$direct(x, y)}: Direct product by pasting together the \code{i}th element
of \code{x} with that of \code{y}.
\item \verb{$inner(x, y)}: Inner product by collapsing the direct product of \code{x}
and \code{y} into a single string.
}
}

\subsection{Method Arguments}{
\itemize{
\item \code{x}: Character vector
\item \code{y}: Character vector
}
}
}
\description{
Create an object for multiplying label vectors together
}
\examples{
multiplier = LabelMultiply("_", ", ")
sir = c("S", "I", "R")
vax = c("unvax", "vax")
multiplier$outer(sir, vax)
multiplier$direct(sir, sir)
multiplier$inner(sir, sir)
}
