% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_products.R
\name{LabelWrapper}
\alias{LabelWrapper}
\title{Label Wrapper}
\usage{
LabelWrapper(open = "(", close = ")")
}
\arguments{
\item{open}{Openning parenthesis symbol.}

\item{close}{Closing parenthesis symbol.}
}
\description{
Construct an object with a \code{wrap} method, which parenthetically wraps
the elements of a string vector.
}
