% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{Ledger}
\alias{Ledger}
\title{Ledger}
\usage{
Ledger(frame, column_map, reference_index_list, labelling_column_names_list)
}
\description{
Make an object to describe ledgers between the entities in an
\code{\link{Index}}. \code{Ledger} object are created by operating on existing
\code{\link{Index}} objects. For example, here the \code{\link{mp_join}}
combines two

\if{html}{\out{<div class="sourceCode {r}">}}\preformatted{age = mp_index(Age = c("young", "old"))
state = mp_cartesian(
  mp_index(Epi = c("S", "I", "R")),
  age
)
mp_join(
  from = mp_subset(state, Epi = "S"),
  to = mp_subset(state, Epi = "I"),
  from.to = "Age"
)
}\if{html}{\out{</div>}}
}
