% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_join.R
\name{LedgerDefinition}
\alias{LedgerDefinition}
\title{Ledgers}
\description{
A ledger is a table with rows that identify specific instances of a
functional form used to define a \code{\link{mp_dynamic_model}}. Ledgers
are most commonly created using the \code{\link{mp_join}} function as in the
following example.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{age = mp_index(Age = c("young", "old"))
state = mp_cartesian(
  mp_index(Epi = c("S", "I", "R")),
  age
)
mp_join(
  from = mp_subset(state, Epi = "S"),
  to = mp_subset(state, Epi = "I"),
  by = list(from.to = "Age")
)
#>     from      to
#>  S.young I.young
#>    S.old   I.old
}\if{html}{\out{</div>}}
}
\concept{population-structure}
