% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{Link}
\alias{Link}
\title{Link}
\usage{
Link(frame, column_map, reference_index_list, labelling_names_list)
}
\description{
Make an object to describe links between the entities in an
\code{\link{Index}}. \code{Link} object are created by operating on existing
\code{\link{Index}} objects. For example, here the \code{\link{mp_join}}
combines two

\if{html}{\out{<div class="sourceCode r">}}\preformatted{age = mp_index(Age = c("young", "old"))
state = mp_cartesian(
  mp_index(Epi = c("S", "I", "R")),
  age
)
mp_join(
  mp_choose(state, "from", Epi = "S"),
  mp_choose(state, "to",   Epi = "I"),
  from.to = "Age"
)
#>     from      to
#>    S.old   I.old
#>  S.young I.young
}\if{html}{\out{</div>}}
}
