% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macpan2.R
\docType{package}
\name{macpan2-package}
\alias{macpan2}
\alias{macpan2-package}
\title{\pkg{macpan2}}
\description{
This R package is a do-over of the
\href{https://github.com/mac-theobio/mcmasterpandemic}{McMasterPandemic}
package on compartmental epidemic modelling for
forecasting and analysis of infectious diseases.
}
\author{
\strong{Maintainer}: Steve Walker \email{swalk@mcmaster.ca}

Authors:
\itemize{
  \item Darren Flynn-Primrose
  \item Weiguang Guan
  \item Steve Walker
}

}
