% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_expr.R
\name{make_expr_parser}
\alias{make_expr_parser}
\title{Generate an Arithmetic Expression Parser}
\usage{
make_expr_parser(parser_name = "parse_expr", finalizer = force)
}
\arguments{
\item{parser_name}{character vector with names of valid functions}

\item{valid_vars}{character vector with names of valid variables}
}
\description{
Generate an Arithmetic Expression Parser
}
