% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{MathExpression}
\alias{MathExpression}
\alias{MathExpressionFromFunc}
\alias{MathExpressionFromStrings}
\title{Math Expression}
\usage{
MathExpressionFromFunc(math_function)

MathExpressionFromStrings(arguments, string)
}
\arguments{
\item{math_function}{Function that takes any number of named
\code{\link{numeric}} arguments and returns a \code{\link{numeric}}
object. The only functions that can be used in this function include
\code{\link{colSums}}, \code{\link{\%*\%}}, \code{\link{[}}, \code{\link{^}}, \code{\link{(}}, \code{\link{rowSums}}, \code{\link{c}}, \code{\link{*}}, \code{\link{sum}}, \code{\link{matrix}}, \code{\link{+}}, \code{\link{-}}, \code{\link{/}}, \code{\link{rep}}.}

\item{arguments}{Character vector of the names of the arguments of
the mathematical expression.}

\item{string}{String representing the expression, referring only to
the names of the arguments and the names of the following functions:
\code{\link{colSums}}, \code{\link{\%*\%}}, \code{\link{[}}, \code{\link{^}}, \code{\link{(}}, \code{\link{rowSums}}, \code{\link{c}}, \code{\link{*}}, \code{\link{sum}}, \code{\link{matrix}}, \code{\link{+}}, \code{\link{-}}, \code{\link{/}}, \code{\link{rep}}.}
}
\description{
Construct objects for representing mathematical expressions.
These objects allow for both numeric and symbolic
evaluations of the expressions. Only the following functions
can be used to construct mathematical expressions:
\code{\link{colSums}}, \code{\link{\%*\%}}, \code{\link{[}}, \code{\link{^}}, \code{\link{(}}, \code{\link{rowSums}}, \code{\link{c}}, \code{\link{*}}, \code{\link{sum}}, \code{\link{matrix}}, \code{\link{+}}, \code{\link{-}}, \code{\link{/}}, \code{\link{rep}}. There are two alternative
constructors. \code{MathExpressionFromFunc} takes a numerical
function and returns a \code{MathExpression} object.
\code{MathExpressionFromString} takes a character vector giving
the names of the arguments and a string representing the expression,
and returns a \code{MathExpression} object.
}
\section{Fields}{

\describe{
\item{\code{string}}{String representing the mathematical expression.}

\item{\code{arguments}}{Character vector of the names of the arguments of
the mathematical expression.}
}}

\section{Methods}{

\itemize{
\item \verb{$numeric_function()}: Function to numerically evaluate the mathematical
expression.
\item \verb{$symbolic_function()}: Function to symbolically evaluate the
mathematical expression.
}
\subsection{Method Arguments}{

The argument signature depends on the expression itself,
and can be found in the \code{arguments} field.
}
}

