% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_model.R
\name{MatsList}
\alias{MatsList}
\title{Matrix List}
\usage{
MatsList(
  ...,
  .mats_to_save = character(0L),
  .mats_to_return = character(0L),
  .dimnames = list(),
  .structure_labels = NullLabels(),
  .init_saved_dims = list()
)
}
\arguments{
\item{...}{Named objects that can be coerced to numerical matrices.}

\item{.mats_to_save}{Character vector naming matrices to be saved at each
set in the simulation so that some calculations can make use of past value
(e.g. delayed effects) and/or to be able to retrieved the simulation
history after the simulation is complete.}

\item{.mats_to_return}{Character vector naming matrices to be returned
after the simulate is complete.}

\item{.dimnames}{Named list of \code{\link{dimnames}} for matrices that change
their dimensions over the simulation steps. These names correspond to the
names of the matrices. The output of the simulations will try their best
to honor these names, but if the shape of the matrix is too inconsistent
with the \code{\link{dimnames}} then numerical indices will be used instead.
For matrices that do not change their dimensions, set \code{\link{dimnames}}
by adding \code{\link{dimnames}} to the matrices passed to \code{...}.}

\item{.structure_labels}{An optional object for obtaining labels of
elements of special vectors and matrices. Such an object can be found in
the \verb{$labels} field of a \code{\link{Compartmental}} model. Note that this
is an advanced technique.}
}
\value{
Object of class \code{MatsList} with the following methods.
\subsection{Methods}{
\itemize{
\item \verb{$data_arg()}: Return the following components of the data structure
to pass to C++.
\itemize{
\item \code{mats} -- Unnamed list of numeric matrices.
\item \code{mats_save_hist} -- Boolean vector identifying which matrices should
have their history saved.
\item \code{mats_return} -- Boolean vector identifying which matrices should be
returned after a simulation.
}
\item \verb{$mat_dims()}: Return a data frame giving the numbers of rows and columns
of each matrix in the list.
\item \verb{$add_mats(...)}: Add matrices to the list and return a new
regenerated \code{MatsList} object.
}
}
}
\description{
Create a list of initial values for matrices used to define a compartmental
model in TMB.
}
