% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_methods.R
\name{MethListFromEngineMethods}
\alias{MethListFromEngineMethods}
\title{Construct a Method List from Engine Methods}
\usage{
MethListFromEngineMethods(engine_methods)
}
\arguments{
\item{engine_methods}{An object of class EngineMethods}
}
\description{
Alternative constuctor for \code{MethList}.
}
\details{
The engine interacts with \code{MethList} objects, but the user interacts
with \code{EngineMethods} objects. This allows developers to convert between
the two.
}
