% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_methods.R
\name{Method}
\alias{Method}
\title{Engine Method Class}
\usage{
Method(name, mat_args, const_args)
}
\arguments{
\item{name}{Method name.}

\item{mat_args}{List of names of matrices that will be used by the method
to produce an output matrix.}

\item{const_args}{List of names of constants that will be used by the method
to produce an output matrix.}
}
\description{
Engine methods allow users to interact with the engine
by specifying a unique \code{name}. This name is used to reference recipes
for computing matrices from other matrices and constants.
}
