% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_methods.R
\name{MethodPrototype}
\alias{MethodPrototype}
\title{Method Prototype}
\usage{
MethodPrototype(formula, mat_arg_nms, int_vec_arg_nms)
}
\arguments{
\item{formula}{Formula for defining a method type using a prototype.}

\item{mat_arg_nms}{Character vector naming the matrix-valued arguments.}

\item{int_vec_arg_nms}{Character vector naming the integer-vector-valued
arguments.}
}
\description{
Define a method type using a prototype. These prototypes can be compared
with methods defined in R to see if they are consistent with a method
type that has been defined in C++. All
arguments are automatically derived through comments in the C++ code where
the method types are defined.
}
