% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_methods.R
\name{MethodTypeUtils}
\alias{MethodTypeUtils}
\title{Method Type Utilities}
\usage{
MethodTypeUtils()
}
\description{
This class is here so that \code{MethodTypes}, which is automatically generated
from the C++ code, can inherit from it.
}
