% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_methods.R
\name{mk_meth_cls}
\alias{mk_meth_cls}
\title{Make Method Class}
\usage{
mk_meth_cls(cls_nm, meth_type_id)
}
\arguments{
\item{cls_nm}{Character string giving the name of the class.}

\item{meth_type_id}{Integer giving the associated ID of the method type.}
}
\description{
Place a method object in the package namespace for a given method type
defined in the C++ code.
}
