% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data_structure.R
\name{Model}
\alias{Model}
\title{Model}
\usage{
Model(definition)
}
\arguments{
\item{definition}{Output of \code{\link{ModelFiles}}.}
}
\description{
Construct an object for representing a model structure.
}
