% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_library.R
\name{model_starter}
\alias{model_starter}
\title{Model Starter}
\usage{
model_starter(starter_name, dir)
}
\arguments{
\item{starter_name}{Currently can only be \code{sir}.}

\item{dir}{String giving the path to a directory for copying the
template model definition.}
}
\description{
Create a directory with a template model definition.
}
