% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data_structure.R
\name{ModelCollection}
\alias{ModelCollection}
\title{Model Collection}
\usage{
ModelCollection(variables, derivations, flows, settings)
}
\arguments{
\item{variables}{Return value of the \code{variables} method in a
\code{\link{ModelFiles}} object.}

\item{derivations}{Return value of the \code{derivations} method in a
\code{\link{ModelFiles}} object.}

\item{flows}{Return value of the \code{flows} method in a
\code{\link{ModelFiles}} object.}

\item{settings}{Return value of the \code{settings} method in a
\code{\link{ModelFiles}} object.}
}
\description{
A model definition that is untied from a set of \code{\link{ModelFiles}}.
}
