% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_files.R
\name{ModelFiles}
\alias{ModelFiles}
\title{Model Files}
\usage{
ModelFiles(
  model_directory,
  csv_reader = CSVReader,
  json_reader = JSONReader,
  txt_reader = TXTReader
)
}
\arguments{
\item{model_directory}{String giving a path to a directory containing
the following files, \code{variables.csv}, \code{derivations.json}, \code{flows.csv},
and \code{settings.json}, described by
\href{https://canmod.net/misc/model_definitions}{this spec}.}

\item{csv_reader}{Class inheriting from \code{\link{Reader}} for reading
csv files.}

\item{json_reader}{Class inheriting from \code{\link{Reader}} for reading
json files.}

\item{txt_reader}{Class inheriting from \code{\link{Reader}} for reading
txt files.}
}
\description{
Construct objects for accessing and caching model definition files.
}
\examples{
d = system.file("starter_models", "seir_symp_vax", package = "macpan2")
m = ModelFiles(d)
m$flows()
expander = FlowExpander(m)
expander$expand_flows()

}
