% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-plot-utils.R
\name{mp_adjacency}
\alias{mp_adjacency}
\title{Adjacency Matrix}
\usage{
mp_adjacency(spec, include_inout = FALSE)
}
\arguments{
\item{spec}{A model specification object (for example, created using
\code{mp_tmb_model_spec()}).}

\item{include_inout}{(logical) include nodes defined by
\code{\link{mp_per_capita_inflow}} and \code{\link{mp_per_capita_outflow}}?}
}
\value{
An adjacency \code{\link{matrix}}.
}
\description{
Get the \href{https://en.wikipedia.org/wiki/Adjacency_matrix}{adjacency matrix}
associated with a model specification.
}
