% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_operator.R
\name{BinaryOperator}
\alias{BinaryOperator}
\alias{mp_binary_operator}
\title{Binary Operator}
\usage{
BinaryOperator(operator)

mp_binary_operator(operator)
}
\arguments{
\item{operator}{A binary operator. Valid binary operations have the following characteristics. \cr - They are functions. \cr - They have exactly two arguments. \cr - They do not have any ... arguments}
}
\value{
A binary operator consistent with the \code{C++} engine.
}
\description{
Convert a function that represents an elementwise binary
operator into one that is consistent with the \code{C++}
engine. This function is intended to clarify how \pkg{macpan2}
treats binary operators, which is a little different from
base \proglang{R}. The difference is clarified in
\code{vignette("elementwise_binary_operators")}, and \code{BinaryOperator} is
primarily used as a resource for that vignette.
}
\section{Functions}{
\itemize{
\item \code{BinaryOperator()}: Synonym of \code{mp_binary_operator}.

}}
\examples{
set.seed(1L)
A = matrix(abs(rnorm(6)), 3, 2)  # 3 by 2 matrix
x = matrix(abs(rnorm(3)))        # 3 by 1 matrix
y = t(abs(rnorm(2)))             # 1 by 2 matrix
times = BinaryOperator(`*`)
pow = BinaryOperator(`^`)
identical(times(A, x), times(x, A))  ## TRUE
identical(pow(A, y), pow(y, A))  ## FALSE

}
