% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{mp_cartesian}
\alias{mp_cartesian}
\title{Cartesian Product of Indexes}
\usage{
mp_cartesian(x, y)
}
\arguments{
\item{x, y}{Objects produced by \code{\link{mp_index}} or derived
from such an object using one of (TODO: list the functions that
will preserve indexness).}
}
\description{
Produce a new index by taking all possible pairwise combinations
of the input indexes. This is useful for producing product models
that expand model components through stratification.
}
\examples{
mp_cartesian(
  mp_index(Epi = c("S", "I")),
  mp_index(Age = c("young", "old"))
)

si = mp_index(Epi = c("S", "I"))
age = mp_index(Age = c("young", "old"))
loc = mp_index(City = c("hamilton", "toronto"))
vax = mp_index(Vax = c("unvax", "vax"))
(si
  |> mp_cartesian(age)
  |> mp_cartesian(loc)
  |> mp_cartesian(vax)
)

flow_rates = mp_index(Epi = c("infection", "recovery"))
mp_union(
  mp_cartesian(
    mp_subset(flow_rates, Epi = "infection"),
    age
  ),
  mp_subset(flow_rates, Epi = "recovery")
)

}
