% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_frame.R
\name{mp_change_frame}
\alias{mp_change_frame}
\title{Change Frame}
\usage{
mp_change_frame(spec)
}
\arguments{
\item{spec}{Model specification (\code{\link{mp_tmb_model_spec}}).}
}
\value{
Data frame with two columns: \code{state} and \code{change}. Each row
describes one change.
}
\description{
Get the changes made to each state variable at each time step.
}
