% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_dynamic_model.R
\name{mp_dynamic_model}
\alias{mp_dynamic_model}
\title{Dynamic Model}
\usage{
mp_dynamic_model(
  expr_list = ExprList(),
  ledgers = list(),
  init_vecs = list(),
  unstruc_mats = list()
)
}
\arguments{
\item{expr_list}{Expression list.}

\item{ledgers}{Ledgers.}

\item{init_vecs}{Initial structured vectors.}

\item{unstruc_mats}{Initial unstructured matrices.}
}
\description{
This is an 'old' model specification function that was tested out
at a workshop. Currently it still drives the engine-agnostic-grammar
vignette, but we plan to replace this function with
\code{\link{mp_tmb_model_spec}} and other model specification
functions.
}
