% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_index.R
\name{mp_extract}
\alias{mp_extract}
\title{Extract Index}
\usage{
mp_extract(x, dimension_name)
}
\arguments{
\item{x}{Object}

\item{dimension_name}{Name of a dimension used in a ledger.}
}
\description{
Extract the index for a particular dimension in a ledger
from a ledger or from an object containing one or more ledgers.
}
