% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{mp_extract_exprs}
\alias{mp_extract_exprs}
\title{Extract Expressions by Name}
\usage{
mp_extract_exprs(x, ...)
}
\arguments{
\item{x}{Object containing named expressions.}

\item{...}{Character vectors containing names of expressions.}
}
\value{
An object the same type as \code{x} but only with
those expressions identified by the names in \code{...}.
}
\description{
Extract Expressions by Name
}
