% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_bulk_indices.R
\name{mp_factors}
\alias{mp_factors}
\title{Factor an Index}
\usage{
mp_factors(index, unpack = c("no", "maybe", "yes"))
}
\arguments{
\item{index}{An index to be factored.}

\item{unpack}{Place factors in the global environment?}
}
\description{
Factor an Index
}
