% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_model.R
\name{mp_final}
\alias{mp_final}
\alias{mp_final_list}
\title{Final Values}
\usage{
mp_final(model)

mp_final_list(model)
}
\arguments{
\item{model}{Object that can be used to simulate.}
}
\description{
Return the values of variables after the simulation loop has finished
and the \code{final} set of expressions have been evaluated.
}
\section{Functions}{
\itemize{
\item \code{mp_final_list()}: Final values formatted as a list of matrices.

}}
