% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_methods.R
\name{mp_fixed_effects}
\alias{mp_fixed_effects}
\title{Fixed Effects}
\usage{
mp_fixed_effects(model)
}
\arguments{
\item{model}{Object that contains information about fitted fixed effects}
}
\value{
A data frame that describes the fitted fixed effects.
}
\description{
Fixed Effects
}
