% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_frame.R
\name{mp_flow_frame}
\alias{mp_flow_frame}
\title{Flow Frame (experimental)}
\usage{
mp_flow_frame(spec, topological_sort = TRUE, warn_not_dag = TRUE)
}
\arguments{
\item{spec}{A \code{\link{mp_tmb_model_spec}}.}

\item{topological_sort}{Should the states be topologically sorted to
respect the main direction of flow?}

\item{warn_not_dag}{Should a warning be thrown if the flow model is
not a directed acyclic graph. This is only relevant if \code{topological_sort}
is used.}
}
\value{
A data frame that gives information provided in calls to
\code{\link{mp_per_capita_flow}} and \code{\link{mp_per_capita_inflow}}.
}
\description{
Get a data frame representing the flows in a model specification.
}
