% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_forecaster.R
\name{mp_forecaster}
\alias{mp_forecaster}
\title{Make a Forecaster}
\usage{
mp_forecaster(
  calibrator,
  forecast_period_time_steps,
  outputs = NULL,
  data = NULL,
  default = list()
)
}
\arguments{
\item{calibrator}{Object made using \code{\link{mp_tmb_calibrator}}, which
can be calibrated to data.}

\item{forecast_period_time_steps}{The number of time steps to project
beyond the period with data.}

\item{outputs}{An optional character vector of variables in the model
to forecast. If this argument is omited, then the forecaster will inherit
the outputs of the \code{calibrator} object. Note that if you prefix the name
of an output using \code{log_}, \code{logit_}, or \code{sqrt_} then the transformed
version of the outputs will be simulated. This technique is useful for
confidence intervals produced by \code{\link{mp_trajectory_sd}} that
go out of the valid range of values (e.g., use \code{log_} for confidence
intervals of negative state variables).}

\item{data}{An optional data frame containing the data that were fitted to.
Typically this argument will be unused, because by default the data are
stored in the \code{calibrator} unless you want to avoid making too many copies
of the data.}

\item{default}{An optional list of default model variables (e.g., parameters
initial values of state variables) to override calibrated values.}
}
\description{
A forecaster is an object that can make forecasts. It is constructed from
\code{\link{mp_tmb_calibrator}}, which is an object that can be used to
calibrate the parameters of a model -- typically by fitting the model to
data. The main difference between a calibrator and a forecaster is that
the latter runs for more time steps, past the last time step of the
calibrator. A forecaster object can be used to generate different types of
forecasts using the \code{\link{mp_trajectory}} family of functions.
}
