% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_data.R
\name{mp_formula_data}
\alias{mp_formula_data}
\title{#' @export
print.FormulaData = function(x, ...) {
print(x$frame, row.names = FALSE)
}}
\usage{
mp_formula_data(...)
}
\description{
#' @export
print.FormulaData = function(x, ...) {
print(x$frame, row.names = FALSE)
}
}
