% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_list.R
\name{mp_functions_used}
\alias{mp_functions_used}
\alias{mp_generates_randomness}
\title{Functions Used by an Object for Communicating with a Computational Engine}
\usage{
mp_functions_used(object)

mp_generates_randomness(object)
}
\arguments{
\item{object}{An object for communicating with a computational engine.}
}
\value{
Character vector of names of functions that are used by \code{object}.
}
\description{
Functions Used by an Object for Communicating with a Computational Engine
}
\section{Functions}{
\itemize{
\item \code{mp_generates_randomness()}: Does the object use functions that generate
randomness?

}}
