% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_group.R
\name{mp_group}
\alias{mp_group}
\title{Group an Index}
\usage{
mp_group(index, by)
}
\arguments{
\item{index}{Index to group rows.}

\item{by}{Column set label to group by.}
}
\description{
Create a new index with fewer columns to create names for
an aggregated vector that is labelled by the input index.
}
