% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\docType{data}
\name{mp_identity}
\alias{mp_identity}
\title{Identity Transformation}
\format{
An object of class \code{DistrParamIdentity} (inherits from \code{DistrParamTrans}, \code{Base}) of length 4.
}
\usage{
mp_identity
}
\value{
DistrParamTrans
}
\description{
Identity transformation for distributional parameters of class \code{DistrParam}.
}
\keyword{datasets}
