% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_model.R
\name{mp_initial}
\alias{mp_initial}
\alias{mp_initial_list}
\title{Initial Values}
\usage{
mp_initial(model)

mp_initial_list(model)
}
\arguments{
\item{model}{A model object from which to extract initial values.}
}
\value{
A long-format data frame with initial values for matrices.
The columns of this output are \code{matrix}, \code{row}, \code{col}, and \code{value}.
Scalar matrices do not have any entries in the \code{row} or \code{col} columns.
}
\description{
Return a data frame containing the values of variables at the end of the
\code{before} phase, right before the simulation loop begins (i.e. right before
the \code{during} phase).
}
\section{Functions}{
\itemize{
\item \code{mp_initial_list()}: List of the initial variables as matrices.

}}
