% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_operator.R
\name{mp_kronecker_operator}
\alias{mp_kronecker_operator}
\title{Kronecker Operator}
\usage{
mp_kronecker_operator(operator)
}
\arguments{
\item{operator}{A scalar binary operator.}
}
\value{
A Kronecker operator convenient for use with \code{macpan2}.
}
\description{
Convert a scalar binary operator into one that performs a
Kronecker product with more convenient row and column names
for use with \code{macpan2}. The result is numerically identical
to base \code{R}'s \code{\%x\%} and \code{kronecker()}, but with cleaner naming
of rows and columns.
}
