% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_labels.R
\name{mp_labels}
\alias{mp_labels}
\title{Index Labels}
\usage{
mp_labels(x, labelling_column_names)
}
\arguments{
\item{x}{Object}

\item{labelling_column_names}{What index columns should be used
for generating the labels. If missing then defaults will be used.
(FIXME: clarify how the defaults are used.)}
}
\description{
Return a character vector of labels for each row of an index
(or a ledger?? FIXME: what does this mean for ledgers??).
}
