% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-plot-utils.R
\name{mp_layout_grid}
\alias{mp_layout_grid}
\title{Flow Diagram Grid Layout}
\usage{
mp_layout_grid(
  spec,
  east = "",
  south = "^$",
  north = "^$",
  west = "^$",
  loops = north,
  x_gap = 0.3,
  y_gap = 0.3,
  north_south_sep = 0,
  east_west_sep = 0,
  trim_states = character()
)
}
\arguments{
\item{spec}{A model specification made with \code{\link{mp_tmb_model_spec}}
or related function.}

\item{east}{Regular expression for matching the names of flows that will
be connected eastward in the layout.}

\item{south}{Regular expression for matching the names of flows that will
be connected southward in the layout.}

\item{north}{Regular expression for matching the names of flows that will
be connected northward in the layout.}

\item{west}{Regular expression for matching the names of flows that will
be connected westward in the layout.}

\item{loops}{Regular expression for matching the names of flows that cause
loops in the flow model, and so should be ignored when building the layout.}

\item{x_gap}{Size of the gap to the left and right of the 1-by-1 space
provided for a node.}

\item{y_gap}{Size of the gap above and below the 1-by-1 space
provided for a node.}

\item{north_south_sep}{Horizontal separation between north and south
flow arrows.}

\item{east_west_sep}{Vertical separation between east and west flow arrows.}

\item{trim_states}{List of states to remove from the diagram}
}
\description{
Create a grid on which to layout the flow diagram of a model specification.
}
