% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-plot-utils.R
\name{mp_layout_paths}
\alias{mp_layout_paths}
\title{Flow Diagram Grid Layout}
\usage{
mp_layout_paths(
  spec,
  sort_paths = TRUE,
  combine_columns = TRUE,
  deduplicate_edges = TRUE,
  loops = "^$",
  ignore = "^$",
  x_gap = 0.3,
  y_gap = 0.3,
  north_south_sep = 0,
  east_west_sep = 0
)
}
\arguments{
\item{spec}{A model specification made with \code{\link{mp_tmb_model_spec}}
or related function.}

\item{sort_paths}{Should the paths/rows be sorted to minimize the
number times an edge must go through a node that it is not connected with?}

\item{combine_columns}{Should each state/node get its own column in the
layout (\code{FALSE}) or should the algorithm try to place branching states
in the same column (\code{TRUE}, default).}

\item{deduplicate_edges}{Should each row have all of the edges in the path
or should duplicate edges be removed?}

\item{loops}{Regular expression for matching the names of flows that cause
loops in the flow model, and so should be ignored when building the layout.}

\item{ignore}{Regular expression for matching the names of flows that should
be removed from the layout analysis entirely. These will be isolated in
a data frame for custom drawing of 'difficult' edges.}
}
\description{
Layout the flow diagram of a model specification so that each row is
one of the paths through the model (ignoring loops).
}
