% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_data.R
\name{mp_ledgers}
\alias{mp_ledgers}
\title{Bundle up Ledgers}
\usage{
mp_ledgers(...)
}
\arguments{
\item{...}{Ledgers to bundle up.}
}
\description{
Bundle up several ledgers (see \code{\link{LedgerDefinition}}) to pass
to \code{\link{mp_dynamic_model}}.
}
