% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_cartesian.R
\name{mp_linear}
\alias{mp_linear}
\title{Linear Chain Product}
\usage{
mp_linear(x, y_labelling_column_names)
}
\arguments{
\item{x}{An index.}

\item{y_labelling_column_names}{TODO}
}
\description{
TODO: what does this mean?
}
\seealso{
Other functions that take products of index tables and return one index tables
\code{\link{mp_cartesian}()},
\code{\link{mp_square}()},
\code{\link{mp_symmetric}()},
\code{\link{mp_triangle}()}
}
\concept{products}
