% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\docType{data}
\name{mp_log}
\alias{mp_log}
\title{Log Transformation}
\format{
An object of class \code{DistrParamLog} (inherits from \code{DistrParamTrans}, \code{Base}) of length 4.
}
\usage{
mp_log
}
\value{
DistrParamTrans
}
\description{
Log transformation for distributional parameters of class \code{DistrParam}.
}
\keyword{datasets}
