% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{mp_lookup}
\alias{mp_lookup}
\title{Lookup}
\usage{
mp_lookup(index, symbol)
}
\arguments{
\item{index}{Index table (see \code{\link{mp_index}}).}

\item{symbol}{Character string that could possibly be associated with a
subset or factor of \code{index}.}
}
\description{
Lookup a subset or factor index associated with a symbol, and return the
index associated with that symbol.
}
