% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model_index.R
\name{mp_model_docs}
\alias{mp_model_docs}
\title{Browse Model Docs}
\usage{
mp_model_docs(model_name, macpan_library = "starter_models")
}
\arguments{
\item{model_name}{Name of a model in the \code{macpan_library}.}

\item{macpan_library}{Name of a library. Currently, the default value of
\code{macpan_library = "starter_models"} is the only recommended option.}
}
\value{
This function returns the URL of the library model, but the main
purpose is the side-effect of automatically opening a web browser at this
URL.
}
\description{
Open a browser at the current version of a particular model in an
online \code{macpan2} model library.
}
