% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_library.R
\name{mp_model_starter}
\alias{mp_model_starter}
\title{Copy Existing Model as a Starting Point}
\usage{
mp_model_starter(starter_name, dir)
}
\arguments{
\item{starter_name}{Currently can only be \code{sir}.}

\item{dir}{String giving the path to a directory for copying the
template model definition.}
}
\description{
Create a directory with a template model definition.
}
\concept{create-model-spec}
