% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{mp_nofit}
\alias{mp_nofit}
\title{Not-Fitting Distributional Parameters}
\usage{
mp_nofit(x, trans = DistrParamTransDefault())
}
\arguments{
\item{x}{fixed numeric value for the distributional parameter, or character
name of an existing variable in the model with a default starting
value to use.}

\item{trans}{transformation to apply to the distributional parameter.
By default, distributional parameters inherit a default transformation from
the associated distribution. For example, the standard deviation parameter
\code{sd} in the \code{\link{mp_normal}} distributions has a default
transformation.}
}
\description{
By default, distributional parameters in priors and likelihoods are not fit.
}
