% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{mp_normal}
\alias{mp_normal}
\title{Normal Distributon}
\usage{
mp_normal(location = NULL, sd = 1)
}
\arguments{
\item{location}{Location parameter.}

\item{sd}{Standard deviation parameter.}
}
\description{
Normal Distributon
}
