% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{mp_opt_attempted}
\alias{mp_opt_attempted}
\title{Optimization Attempted}
\usage{
mp_opt_attempted(model)
}
\arguments{
\item{model}{A model that can be calibrated, probably produced using
\code{\link{mp_tmb_calibrator}}.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
Has an attempt been made to calibrate model parameters through optimization
of a likelihood function or posterior density, probably by using
\code{\link{mp_optimize}}?
}
