% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_tmb_calibrator.R
\name{mp_optimize}
\alias{mp_optimize}
\alias{mp_optimize.TMBSimulator}
\title{Optimize}
\usage{
mp_optimize(model, optimizer, ...)

\method{mp_optimize}{TMBSimulator}(model, optimizer = c("nlminb", "optim"), ...)
}
\arguments{
\item{model}{A model object capable of being optimized. See below
for model types that are supported.}

\item{optimizer}{Name of an implemented optimizer. See below for options
for each type of \code{model}.}

\item{...}{Arguments to pass to the \code{optimizer}.}
}
\value{
The output of the \code{optimizer}. The \code{model} object is modified
and saves the history of optimization outputs (TODO: develop a way for
the user to get a hold of this stuff).
}
\description{
Optimize
}
\section{Methods (by class)}{
\itemize{
\item \code{mp_optimize(TMBSimulator)}: Optimize a TMB simulator. TODO: less technical ...

}}
