% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_tmb_calibrator.R
\name{mp_optimized_spec}
\alias{mp_optimized_spec}
\title{Optimized Model Specification}
\usage{
mp_optimized_spec(model, spec_structure = c("original", "modified"))
}
\arguments{
\item{model}{A model object that can be optimized/calibrated. Currently,
only models produced by \code{\link{mp_tmb_calibrator}} are valid.}

\item{spec_structure}{A character string identifying the structure of the
returned model specification. Use \code{"original"} for the specification
originally passed to the \code{\link{mp_tmb_calibrator}} function, and
\code{"modified"} for the one that was used to fit to data. See the details
below for more information on these modifications.}
}
\value{
A copy of the model specification used to produce the calibrator
model, with calibrated default values.
}
\description{
Create a new model specification using parameter values that have been
optimized. See the examples in the help page for
\code{\link{mp_tmb_calibrator}}.
}
\details{
The following is a list of additional model specification structure that was
or could have been added to the \code{"modified"} specification by
\code{\link{mp_tmb_calibrator}} for the purposes of fitting to data.
\itemize{
\item Observed data in a format that can be compared with simulated data.
\item Expressions preparing simulated data so that they can be compared with the
observed data.
\item Distributional parameters for likelihood and prior components of the
objective function (e.g., a dispersion parameter for a negatively binomial
likelihood component, or a prior standard deviation for a transmission rate
parameter).
\item Data, parameters, and expressions for modelling time-variation of
parameters (e.g., basis functions and weights for a spline determining the
time-variation of transmission rate).
}
}
