% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_tmb_calibrator.R
\name{mp_optimizer_output}
\alias{mp_optimizer_output}
\title{Optimizer Output}
\usage{
mp_optimizer_output(model, what = c("latest", "all"))
}
\arguments{
\item{model}{An object that has been optimized.}

\item{what}{A string indicating whether to return the results of the
\code{"latest"} optimization attempt or a list with \code{"all"} of them.}
}
\description{
Get the output from an optimizer used in model calibration.
}
\details{
When objects created by \code{\link{mp_tmb_calibrator}} are successfully
passed to \code{\link{mp_optimize}}, they build up an
optimization history. This history is recorded as a list of the output
produced by the underlying optimizer (e.g. \code{\link{nlminb}}). This
\code{mp_optimizer_output} function returns the latest output by default
or the entire history list.
}
