% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrator_arg_constructors.R
\name{mp_par}
\alias{mp_par}
\title{Fit Parameters}
\usage{
mp_par(params, random)
}
\arguments{
\item{params}{Named list of distributional specifications for the
fixed effects.}

\item{random}{Named list of distributional specifications for the random
effects.}
}
\description{
Define the prior distributions for parameters and random effects to be
passed to \code{par} argument of the \code{\link{mp_tmb_calibrator}} function.
}
