% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_parameters.R
\name{mp_parameterization}
\alias{mp_parameterization}
\title{Description of Model Parameterization}
\usage{
mp_parameterization(model, types = c("fixed", "random"))
}
\arguments{
\item{model}{Parameterized model, probably produced using
\code{\link{mp_tmb_calibrator}}.}

\item{types}{Vector indicating what kinds of parameters should be
included, \code{"fixed"} for fixed-effect-type parameters and \code{"random"}
for random-effect-type.}
}
\description{
Description of Model Parameterization
}
\examples{
cal = si_example_calibrator()
mp_parameterization(cal)
}
