% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_list_generators.R
\name{mp_per_capita_flow}
\alias{mp_per_capita_flow}
\title{Per-Capita Flow}
\usage{
mp_per_capita_flow(from, to, rate)
}
\arguments{
\item{from}{String giving the name of the compartment from which the flow
originates.}

\item{to}{String giving the name of the compartment to which the flow is
going.}

\item{rate}{Can be one of three formats: (1) a two-sided formula with the
left-hand-side and right-hand-side giving the name and expression for the
per-capita rate of flow from \code{from} to \code{to}, (2) a one-sided formula...
(3) a string giving the name of a variable that contains the per-capita flow.
(Currently only (1) is implemented)}
}
\description{
Per-Capita Flow
}
