% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{mp_poisson}
\alias{mp_poisson}
\title{Poisson Distribution}
\usage{
mp_poisson(location = DistrParam("location"))
}
\arguments{
\item{location}{Location parameter. Only necessary if used as a prior
distribution. If it is used as a likelihood component the location
parameter will be taken as the simulated variable being fitted to data,
and so this \code{location} parameter should be left to the default.}
}
\value{
DistrSpec
}
\description{
Poisson Distribution
}
