% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_index.R
\name{mp_positions}
\alias{mp_positions}
\title{Position Vectors}
\usage{
mp_positions(x, table, zero_based = TRUE)
}
\arguments{
\item{x}{Character vector}

\item{table}{Character vector}

\item{zero_based}{Use zero-based indexing? Defaults to \code{TRUE},
otherwise standard R one-based indexing is used.}
}
\description{
Return an integer vector of positions of \code{x} in
\code{table}. Currently this is a simple wrapper around
\code{\link{match}}.
}
