% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_model.R
\name{mp_print_obj_fn}
\alias{mp_print_obj_fn}
\title{Print Objective Function}
\usage{
mp_print_obj_fn(model)
}
\arguments{
\item{model}{Model object with an objective function, probably a
calibrator produced using \code{\link{mp_tmb_calibrator}}.}
}
\value{
Called to print the objective function for humans to read.
Invisibly returns the underlying objective function object that is
of limited utility.
}
\description{
Print Objective Function
}
