% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_tmb_model_spec.R
\name{mp_print_spec}
\alias{mp_print_spec}
\alias{mp_print_before}
\alias{mp_print_during}
\alias{mp_print_after}
\title{Print Model Specification}
\usage{
mp_print_spec(model)

mp_print_before(model)

mp_print_during(model)

mp_print_after(model)
}
\arguments{
\item{model}{A model produced by \code{\link{mp_tmb_model_spec}}.}
}
\description{
Print Model Specification
}
\section{Functions}{
\itemize{
\item \code{mp_print_before()}: Print just the expressions executed before the
simulation loop.

\item \code{mp_print_during()}: Print just the expressions executed during each
iteration of the simulation loop.

\item \code{mp_print_after()}: Print just the expressions executed after the
simulation loop.

}}
