% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_tmb_model_spec.R
\name{mp_read_rds}
\alias{mp_read_rds}
\title{Read Serialized Model Specification}
\usage{
mp_read_rds(filename)
}
\arguments{
\item{filename}{Path to a saved model specification object.}
}
\description{
Read Serialized Model Specification
}
