% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_list_generators.R
\name{mp_reduce}
\alias{mp_reduce}
\title{Reduce Model}
\usage{
mp_reduce(model)
}
\arguments{
\item{model}{A model object.}
}
\description{
Reduce a model by removing any model structure
(e.g. \code{\link{mp_per_capita_flow}}), so that expression lists
are plain R formulas.
}
