% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_index.R
\name{mp_reference}
\alias{mp_reference}
\title{Reference Index}
\usage{
mp_reference(x, dimension_name)
}
\arguments{
\item{x}{Object}

\item{dimension_name}{Name of a dimension used in a ledger if
applicable.}
}
\description{
Extract the index used as a reference for generating position
vectors.
}
