% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_index.R
\name{mp_rename}
\alias{mp_rename}
\title{Rename Index Columns}
\usage{
mp_rename(x, ...)
}
\arguments{
\item{x}{An index with columns to be renamed.}

\item{...}{Name-value pairs. The name gives the new name and the value
is a character vector giving the old name.}
}
\description{
Rename Index Columns
}
\seealso{
Other functions that return index tables
\code{\link{mp_cartesian}()},
\code{\link{mp_index}()},
\code{\link{mp_subset}()},
\code{\link{mp_union}()}
}
\concept{indexes}
\concept{population-structure}
