% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_to_tmb.R
\name{mp_report}
\alias{mp_report}
\title{reporting function}
\usage{
mp_report(simulator, parameter_vector = numeric(), phases = "during")
}
\arguments{
\item{simulator}{Object produced by \code{\link{tmb_simulator}}.}

\item{parameter_vector}{Numeric vector equal in length to (?)}
}
\description{
reporting function
}
