% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_cal_time.R
\name{mp_sim_bounds}
\alias{mp_sim_bounds}
\title{Simulation Bounds}
\usage{
mp_sim_bounds(sim_start, sim_end, time_scale = "steps", time_column = "time")
}
\arguments{
\item{sim_start}{Start time of each simulation.}

\item{sim_end}{End time of each simulation.}

\item{time_scale}{Qualitative description of the size of a time step.
The only valid values are 'steps' and 'daily'. If you would like each time
step in the model to represent one day, then you should consider using
'daily'. Note that using 'daily' will require that your data represent
time using a (1) \code{\link{Date}} vector, (2) \code{\link{character}}
vector in YYYY-MM-DD format, or (3) \code{\link{integer}} vector that
counts the number of days since some reference. Otherwise please choose
'steps', the default, and convert your time column into integer values that
represent the time step that you would like in the model.}

\item{time_column}{Name of the column that will identify the time at which
particular values were observed.}
}
\value{
An object to be passed to the \code{time} argument of
\code{\link{mp_tmb_calibrator}}.
}
\description{
Set the simulation bounds (start time and end time) for a
calibration. This is used to override the default simulation
bounds taken from the observed data passed to
\code{\link{mp_tmb_calibrator}}.
}
\seealso{
\code{\link[=mp_sim_offset]{mp_sim_offset()}}
}
