% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_cal_time.R
\name{mp_sim_offset}
\alias{mp_sim_offset}
\title{Simulation Offsets (Experimental)}
\usage{
mp_sim_offset(
  sim_start_offset,
  sim_end_offset,
  time_scale,
  time_column = "time"
)
}
\arguments{
\item{sim_start_offset}{Number of time steps before the first data point
to start each simulation.}

\item{sim_end_offset}{Number of time steps after the last data point to
end each simulation.}

\item{time_scale}{Qualitative description of the size of a time step.
currently only \code{"steps"}, \code{"daily"}, and \code{"weekly"} are allowed,
and but \code{"steps"} is the only recommended version as the other two
are poorly tested and will throw a warning. The recommended \code{"steps"}
option assumes that positive integers are used to indicate a particular
point in the simulation.}
}
\description{
Offset the starting and ending times of the simulation, from the
start and end time of the data used in calibration. This is used to
override the default offsets of zero taken from the observed data passed to
\code{\link{mp_tmb_calibrator}}.
}
