% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_bulk_indices.R
\name{mp_slices}
\alias{mp_slices}
\title{Slice an index}
\usage{
mp_slices(index, unpack = c("no", "maybe", "yes"))
}
\arguments{
\item{index}{Index to slice up.}

\item{unpack}{Place factors in the global environment?}
}
\description{
Slice an index
}
