% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_cartesian.R
\name{mp_square}
\alias{mp_square}
\title{Self Cartesian Product}
\usage{
mp_square(x, suffixes = c("A", "B"))
}
\arguments{
\item{x}{A \code{\link{Partition}} object.}

\item{suffixes}{Length-2 character vector giving suffixes that
disambiguate the column names in the output.}
}
\description{
Self Cartesian Product
}
\seealso{
Other functions that take products of index tables and return one index tables
\code{\link{mp_cartesian}()},
\code{\link{mp_linear}()},
\code{\link{mp_symmetric}()},
\code{\link{mp_triangle}()}
}
\concept{products}
