% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_operator.R
\name{mp_square_operator}
\alias{mp_square_operator}
\title{Square Matrix Operator}
\usage{
mp_square_operator(operator)
}
\arguments{
\item{operator}{A unary operator of a vector.}
}
\value{
An operator.
}
\description{
Convert a unary operator that takes a vector, into one with
dimensions named in a way that is more convenient for use with
\code{macpan2}.
}
