% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_frame.R
\name{mp_state_dependence_frame}
\alias{mp_state_dependence_frame}
\title{State Dependence Frame}
\usage{
mp_state_dependence_frame(spec)
}
\arguments{
\item{spec}{Model specification from \code{\link{spec}}.}
}
\description{
Data frame giving states that per-capita flow rates directly depend on.
This is intended for plotting diagrams and not for mathematical analysis,
in that it does not describe indirect dependence for flow rates on state
variables.
}
