% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_frame.R
\name{mp_state_vars}
\alias{mp_state_vars}
\title{State Variables}
\usage{
mp_state_vars(spec)
}
\arguments{
\item{spec}{Model specification (\code{\link{mp_tmb_model_spec}}).}
}
\value{
Character vector of names of all state variables that have been
explicitly represented in the model using functions like
\code{\link{mp_per_capita_flow}}.
}
\description{
Get the state variables in a model specification.
}
