% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured_formulas.R
\name{mp_struc}
\alias{mp_struc}
\title{Structured Formula}
\usage{
mp_struc(formula, by = NULL)
}
\arguments{
\item{formula}{An unstructured formula to be turned into a structured one.}

\item{by}{See \code{\link{mp_join}}.}
}
\description{
Structured Formula
}
