% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{mp_subset}
\alias{mp_subset}
\alias{mp_setdiff}
\title{Subset of Indexes}
\usage{
mp_subset(x, ...)

mp_setdiff(x, ...)
}
\arguments{
\item{x}{Model index.}

\item{...}{Tagged character vectors used to determine the subset that is
taken. The tags refer to columns (or sets of columns using dot-concatenation)
in \code{x} and the values of the character vectors refer to labels with
respect to those columns.}
}
\description{
Take a subset of the rows of an index to produce another index.
}
