% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_cartesian.R
\name{mp_symmetric}
\alias{mp_symmetric}
\title{Symmetric Self Cartesian Product}
\usage{
mp_symmetric(x, y_labelling_column_names, exclude_diag = TRUE)
}
\arguments{
\item{x}{A \code{\link{Partition}} object.}

\item{y_labelling_column_names}{TODO}

\item{exclude_diag}{Should 'diagonal' commponents be excluded from the output.}
}
\description{
Symmetric Self Cartesian Product
}
\seealso{
Other functions that take products of index tables and return one index tables
\code{\link{mp_cartesian}()},
\code{\link{mp_linear}()},
\code{\link{mp_square}()},
\code{\link{mp_triangle}()}
}
\concept{products}
