% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{mp_time_scale}
\alias{mp_time_scale}
\title{Time Scale}
\usage{
mp_time_scale(start, end, time_step_scale = c("steps", "daily", "weekly"), ...)
}
\arguments{
\item{start}{First date or time in the first time step}

\item{end}{Last date or time in the last time step}

\item{time_step_scale}{TODO}

\item{...}{TODO}
}
\description{
Time Scale
}
