% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_tmb_calibrator.R
\name{mp_tmb}
\alias{mp_tmb}
\title{Get Underlying TMB Object}
\usage{
mp_tmb(model)
}
\arguments{
\item{model}{An object based on TMB.}
}
\description{
Get the result of \code{TMB::MakeADFun} underlying a TMB-based
model in \code{macpan2}.
}
