% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_methods.R
\name{mp_tmb_coef}
\alias{mp_tmb_coef}
\title{TMB Model Coefficient Table}
\usage{
mp_tmb_coef(model, ...)
}
\arguments{
\item{model}{Object that contains information about model coefficients.}

\item{...}{Arguments to pass onto the \code{broom.mixed::tidy.TMB} method.}
}
\value{
A data frame that describes the fitted coefficients.
}
\description{
TMB Model Coefficient Table
}
