% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_methods.R
\name{mp_tmb_fixef_cov}
\alias{mp_tmb_fixef_cov}
\title{Covariance of Fixed Effect Estimates}
\usage{
mp_tmb_fixef_cov(model)
}
\arguments{
\item{model}{Object that contains information about fitted model parameters.}
}
\value{
A covariance matrix.
}
\description{
Covariance of Fixed Effect Estimates
}
