% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_model_editors.R
\name{mp_tmb_insert}
\alias{mp_tmb_insert}
\alias{mp_tmb_update}
\alias{mp_tmb_delete}
\title{Modify a TMB Model Spec}
\usage{
mp_tmb_insert(
  model,
  phase = "during",
  at = 1L,
  expressions = list(),
  default = list(),
  integers = list(),
  must_save = character(),
  must_not_save = character(),
  sim_exprs = character()
)

mp_tmb_update(
  model,
  phase = "during",
  at = 1L,
  expressions = list(),
  default = list(),
  integers = list(),
  must_save = character(),
  must_not_save = character(),
  sim_exprs = character()
)

mp_tmb_delete(
  model,
  phase,
  at,
  default = character(),
  integers = character(),
  must_save = character(),
  must_not_save = character(),
  sim_exprs = character()
)
}
\arguments{
\item{model}{TMB model spec object produced using
\code{\link{mp_tmb_library}} or \code{\link{mp_tmb_model_spec}}.}

\item{phase}{At what phase should \code{expressions} be inserted, updated,
or deleted.}

\item{at}{Expression number, which can be identified by printing out
\code{model}, at which the \code{expressions} should be inserted or updated. If
inserted then the existing expressions with number \code{at} and higher are
shifted after the new \code{expressions} are added. If updated, then the
existing expressions with number from \code{at} to \code{at + length(expressions) - 1}
are replaced with the new \code{expressions}.
For \code{mp_tmb_delete}, a numeric vector of integers identifying expressions
to delete from the model.}

\item{expressions}{Expressions to insert into the model spec or to
replace existing expressions.}

\item{default}{Named list of objects, each of which can be coerced into
a \code{\link{numeric}} \code{\link{matrix}}. The names refer to
variables that appear in \code{before}, \code{during}, and \code{after}.
For \code{mp_tmb_delete}, a character vector of such objects to delete from
the model.}

\item{integers}{Named list of vectors that can be coerced to integer
vectors. These integer vectors can be used by name in model formulas to
provide indexing of matrices and as grouping factors in
\code{\link{group_sums}}.
For \code{mp_tmb_delete}, a character vector of such objects to delete from
the model.}

\item{must_save}{Character vector of the names of matrices that must have
their values stored at every iteration of the simulation loop. For example,
a matrix that the user does not want to be returned but that impacts dynamics
with a time lag must be saved and therefore in this list.}

\item{must_not_save}{Character vector of the names of matrices that must
not have their values stored at every iteration of the simulation loop. For
example, the user may ask to return a very large matrix that would create
performance issues if stored at each iteration. The creator of the model
can mark such matrices making it impossible for the user of the model to
save their full simulation history.}

\item{sim_exprs}{Character vector of the names of \code{before},
\code{during}, and \code{after} expressions that must only be evaluated
when simulations are being produced and not when the objective function is
being evaluated. For example, expressions that generate stochasticity should
be listed in \code{sim_exprs} because TMB objective functions must be
continuous.}
}
\value{
A new model spec object with updated and/or inserted information.
}
\description{
Insert, update, or delete elements of a TMB model spec, produced using
\code{\link{mp_tmb_library}} or \code{\link{mp_tmb_model_spec}}, or
\code{\link{mp_tmb_delete}}. The
only difference between \code{mp_tmb_insert} and \code{mp_tmb_update} is that
the former shifts the positions of existing expressions to make room
for the new expressions, whereas the latter overwrites existing expressions
using the new expressions. The treatment of new \code{default} values and
\code{integers} is the same. The examples below clarify this difference.
Note that \code{mp_tmb_delete} does not contain an \code{expressions} argument,
because it is not necessary to specify new expressions in the case
of deletion.
}
\details{
These modifications do not update the model specification in-place. Instead
the output of \code{mp_tmb_insert}, \code{mp_tmb_update}, and \code{mp_tmb_delete} define
a new model specification and should be saved if you want to use the new
model (e.g., \code{new_model = mp_tmb_insert(model, ...)}).
}
\examples{
si = mp_tmb_library("starter_models", "si", package = "macpan2")
print(si)

## Update the mixing process to include 
## optional phenomenological heterogeneity.
## We need mp_tmb_update here so that 
## the previous infection expression is
## overwritten.
mp_tmb_update(si, phase = "during"
  , at = 1
  , expressions = list(infection ~ beta * I * (S/N)^zeta)
  , default = list(zeta = 1)
)

## Parameterize with log_beta in place of beta.
## We need mp_tmb_insert here so that the
## existing expression for computing the initial
## number of susceptible indiviudals is not
## overwritten.
mp_tmb_insert(si, phase = "before"
  , at = 1
  , expressions = list(beta ~ exp(log_beta))
  , default = list(log_beta = log(0.5))
)

}
