% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_model_editors.R
\name{mp_tmb_insert_backtrans}
\alias{mp_tmb_insert_backtrans}
\title{Insert Back Transformations of Model Parameters}
\usage{
mp_tmb_insert_backtrans(
  model,
  variables = character(),
  transformation = mp_log
)
}
\arguments{
\item{model}{TMB model spec object produced using
\code{\link{mp_tmb_library}} or \code{\link{mp_tmb_model_spec}}.}

\item{variables}{Character vector of variables to transform.}

\item{transformation}{A transformation object such as \code{\link{mp_log}},
which is the default. See the help page for \code{\link{mp_log}} for
available options.}

\item{variable}{Character vector of parameters to back transform.}
}
\value{
A new model specification object with expressions for the
untransformed (or back transformed) parameters at the beginning of the
\code{before} phase. The transformed version of the parameter is also
added to the defaults and are identified with a prefixed name (e.g.,
\code{log_beta} if \code{beta} is log transformed).
}
\description{
Insert Back Transformations of Model Parameters
}
\seealso{
\code{\link[=mp_tmb_insert_backtrans]{mp_tmb_insert_backtrans()}}
}
