% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_model_editors.R
\name{mp_tmb_insert_glm_timevar}
\alias{mp_tmb_insert_glm_timevar}
\title{Insert GLM Time Variation}
\usage{
mp_tmb_insert_glm_timevar(
  model,
  parameter_name,
  design_matrix,
  timevar_coef,
  link_function = mp_log,
  matrix_coef_name = sprintf("matrix_coef_\%s", parameter_name),
  matrix_row_name = sprintf("matrix_row_\%s", parameter_name),
  matrix_col_name = sprintf("matrix_col_\%s", parameter_name),
  linear_pred_name = sprintf("linear_pred_\%s", parameter_name),
  timeseries_name = sprintf("timeseries_\%s", parameter_name),
  timevar_coef_name = sprintf("time_var_\%s", parameter_name),
  time_index_name = sprintf("time_index_\%s", parameter_name),
  sparsity_tolerance = 0,
  engine_function = c("sparse_mat_mult", "group_sums")
)
}
\arguments{
\item{model}{A model specification (see \code{\link{mp_tmb_model_spec}}).}

\item{parameter_name}{Character string giving the name of the parameter
to make time-varying.}

\item{design_matrix}{Matrix of time variation.}

\item{timevar_coef}{Initial coefficient matrix with the same number of rows
as there are columns of the \code{design_matrix}.}

\item{link_function}{Link function given by functions like
\code{\link{mp_log}}.}

\item{matrix_coef_name}{Name of the vector containing values of the non-zero
elements of the design matrix.}

\item{matrix_row_name}{Name of the vector containing row indices of
the non-zero elements of the design matrix.}

\item{matrix_col_name}{Name of the vector containing column indices of
the non-zero elements of the design matrix.}

\item{linear_pred_name}{Name of the vector containing the linear
predictor.}

\item{timeseries_name}{Name of the vector containing the time-series
of the varying parameter.}

\item{timevar_coef_name}{Name of the vector containing the time-varying
parameter coefficients.}

\item{time_index_name}{Name of the index at which the time varying
parameter changes.}

\item{sparsity_tolerance}{Make design matrix coefficients exactly zero
when they are below this tolerance.}
}
\description{
Insert GLM Time Variation
}
