% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_model_editors.R
\name{mp_tmb_insert_reports}
\alias{mp_tmb_insert_reports}
\title{Insert Reports}
\usage{
mp_tmb_insert_reports(
  model,
  incidence_name,
  report_prob,
  mean_delay,
  cv_delay,
  reports_name = sprintf("reported_\%s", incidence_name),
  report_prob_name = sprintf("\%s_report_prob", incidence_name),
  mean_delay_name = sprintf("\%s_mean_delay", incidence_name),
  cv_delay_name = sprintf("\%s_cv_delay", incidence_name)
)
}
\arguments{
\item{model}{A model produced by \code{\link{mp_tmb_model_spec}}.}

\item{incidence_name}{Name of the incidence variable to be transformed.}

\item{report_prob}{Value to use for the reporting probability; the
proportion of cases that get reported.}

\item{mean_delay}{Mean of the Gamma distribution of reporting delay times
measured in time steps. For example, if the mean delay time in days is 14
and each time step in the model corresponds to one week, then \code{mean_delay}
should be set to \code{2}.}

\item{cv_delay}{Coefficient of variation of the Gamma distribution of
reporting delay times, with time measured in time steps (see more on this
in the description of the \code{mean_delay} argument immediately above).}

\item{reports_name}{Name of the new reports variable.}

\item{report_prob_name}{Name of the variable containing \code{report_prob}.}

\item{mean_delay_name}{Name of the variable containing \code{mean_delay}.}

\item{cv_delay_name}{Name of the variable containing \code{cv_delay}.}
}
\description{
A version of \code{\link{mp_tmb_insert}} making it more convenient to
transform an incidence variable into a reports variable, which accounts
for reporting delays and under-reporting. This new reports variable is
a convolution of the simulation history of an incidence variable with
a kernel that is proportional to a Gamma distribution of reporting
delay times.
}
\examples{
sir = mp_tmb_library("starter_models", "sir", package = "macpan2")
sir_delays = mp_tmb_insert_reports(sir
  , incidence_name = "infection"
  , report_prob = 0.1
  , mean_delay = 14
  , cv_delay = 0.25
  , reports_name = "reports"
)
sir_delays |> mp_print_during()
(sir_delays 
  |> mp_simulator(
      time_steps = 50L
    , outputs = c("infection", "reports"), 50L)
  |> mp_trajectory()
)

}
