% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_tmb_model_spec.R
\name{mp_tmb_model_spec}
\alias{mp_tmb_model_spec}
\title{Specify a TMB Model}
\usage{
mp_tmb_model_spec(
  before = list(),
  during = list(),
  after = list(),
  default = list(),
  integers = list(),
  must_save = character(),
  must_not_save = character(),
  sim_exprs = character()
)
}
\arguments{
\item{before}{List of formulas to be evaluated (in the order provided) before
the simulation loop begins. Each \code{\link{formula}} must have a left hand
side that gives the name of the matrix being updated, and a right hand side
giving an expression containing only the names of matrices in the model,
functions defined in the TMB engine, and numerical literals (e.g.
\code{3.14}). The available functions in the TMB engine  can be described in
\code{\link{engine_functions}}. Names can be provided for the components of
\code{before}, and these names do not have to be unique. These names are
used by the \code{sim_exprs} argument.}

\item{during}{List of formulas to be evaluated at every iteration of the
simulation loop, with the same rules as \code{before}.}

\item{after}{List of formulas to be evaluated after the simulation loop,
with the same rules as \code{before}.}

\item{default}{Named list of objects that can be coerced into
\code{\link{numeric}} \code{\link{matrices}}. The names refer to
variables that appear in \code{before}, \code{during}, and \code{after}.}

\item{integers}{list}

\item{must_save}{character}

\item{must_not_save}{character}

\item{sim_exprs}{character}
}
\description{
Specify a model in the TMB engine.
}
