% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfplot.R
\name{mp_tmb_profile}
\alias{mp_tmb_profile}
\title{TMB Likelihood Profiling}
\usage{
mp_tmb_profile(model, param, ...)
}
\arguments{
\item{model}{A TMB model probably produced using
\code{\link{mp_tmb_calibrator}}.}

\item{param}{The name of a fixed effect parameter set through the \code{par}
argument of the call to \code{\link{mp_tmb_calibrator}} used to create
\code{model}.}

\item{...}{Arguments to pass to \code{TMB::tmbprofile}.}
}
\value{
The output of \code{TMB::tmbprofile}.
}
\description{
Use \code{TMB::tmbprofile} to compute the profile likelihood
of a calibrator produced using \code{\link{mp_tmb_calibrator}}.
}
