% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_tmb_struc_model.R
\name{mp_tmb_struc_model}
\alias{mp_tmb_struc_model}
\title{Structured TMB Model Specification}
\usage{
mp_tmb_struc_model(
  before = list(),
  during = list(),
  after = list(),
  state_vector = list(),
  other_vectors = list()
)
}
\description{
Structured TMB Model Specification
}
