% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrator_arg_constructors.R
\name{mp_traj}
\alias{mp_traj}
\title{Trajectory Specification}
\usage{
mp_traj(likelihood = empty_named_list(), condensation = empty_named_list())
}
\arguments{
\item{likelihood}{List of likelihood components. The names of the list
identify the trajectory associated with each likelihood component.}

\item{condensation}{List of condensation methods. The names of the list
identify the trajectories produced by each condensation method.}
}
\description{
Specify a set of trajectories to fit. The output of this function
is intended to be passed to the \code{traj} argument of
\code{\link{mp_tmb_calibrator}}.
}
