% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_model.R
\name{mp_trajectory_sim}
\alias{mp_trajectory_sim}
\title{Random Trajectory Simulations}
\usage{
mp_trajectory_sim(model, n, probs = c(0.025, 0.25, 0.5, 0.75, 0.975))
}
\arguments{
\item{model}{Model object that can generate simulations with random numbers.}

\item{n}{Number of random trajectories to simulate.}

\item{probs}{Numeric vector of probabilities corresponding to quantiles for
summarizing the results over the random realizations.}
}
\description{
Random Trajectory Simulations
}
