% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_cartesian.R
\name{mp_triangle}
\alias{mp_triangle}
\title{Self Cartesian Product Excluding One Off-Diagonal Side}
\usage{
mp_triangle(
  x,
  y_labelling_column_names,
  exclude_diag = TRUE,
  lower_tri = FALSE
)
}
\arguments{
\item{x}{An index.}

\item{y_labelling_column_names}{TODO}

\item{exclude_diag}{Should 'diagonal' commponents be excluded from the output.}

\item{lower_tri}{Should the lower triangular components be include from the
output. If \code{FALSE} the result is upper triangular.}
}
\description{
Self Cartesian Product Excluding One Off-Diagonal Side
}
\seealso{
Other functions that take products of index tables and return one index tables
\code{\link{mp_cartesian}()},
\code{\link{mp_linear}()},
\code{\link{mp_square}()},
\code{\link{mp_symmetric}()}
}
\concept{population-structure}
\concept{products}
