% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{mp_uncertainty_estimated}
\alias{mp_uncertainty_estimated}
\title{Uncertainty Estimated}
\usage{
mp_uncertainty_estimated(model)
}
\arguments{
\item{model}{A model that can be calibrated, probably produced using
\code{\link{mp_tmb_calibrator}}.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
Does a model contain estimates of parameter uncertainty, probably
through a covariance matrix estimated using \code{\link{mp_optimize}}?
}
