% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_union.R
\name{mp_union}
\alias{mp_union}
\title{Union of Indexes}
\usage{
mp_union(...)
}
\arguments{
\item{...}{Indexes.}
}
\description{
Union of Indexes
}
\seealso{
Other functions that return index tables
\code{\link{mp_cartesian}()},
\code{\link{mp_index}()},
\code{\link{mp_rename}()},
\code{\link{mp_subset}()}
}
\concept{indexes}
