% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{mp_vector}
\alias{mp_vector}
\title{Stub}
\usage{
mp_vector(x, ...)
}
\description{
This documentation was originally in \code{\link[=mp_index]{mp_index()}} and should be cleaned up
See issue #131
}
\details{
#' These labels can be used to create 'multidimensional' names for the elements
of vectors. Here is the above example expressed in vector form.

\if{html}{\out{<div class="sourceCode {r, echo = FALSE}">}}\preformatted{v = Vector(prod)
v$set_numbers(Epi = c(S = 1000))$set_numbers(Epi = c(I = 1), Age = "old")
}\if{html}{\out{</div>}}

This example vector could be stored as a 3-by-2 matrix. But other examples
cannot, making this indexing approach more general. For example, consider the
following index.

\if{html}{\out{<div class="sourceCode {r, echo = FALSE}">}}\preformatted{}\if{html}{\out{</div>}}

This index has an associated indexed vector that cannot be expressed as a
matrix.

\if{html}{\out{<div class="sourceCode {r, echo = FALSE}">}}\preformatted{mp_vector(symp)$set_numbers(Epi = c(S = 1000))$set_numbers(Epi = c(I = 1), Symptoms = "severe")
}\if{html}{\out{</div>}}
}
\examples{
state = mp_index(
  Epi = c("S", "I", "S", "I"),
  Age = c("young", "young", "old", "old")
)
state_vector = (state
  |> mp_vector()
  |> mp_set_numbers(Epi = c(S = 1000))
  |> mp_set_numbers(Epi = c(I = 1), Age = "old")
)
print(state_vector)

}
