% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macpan2.R
\name{mp_version}
\alias{mp_version}
\title{Version of \code{macpan2}}
\usage{
mp_version(model)
}
\arguments{
\item{model}{\code{macpan2} model object.}
}
\value{
Object of type \code{\link{package_version}} giving the version of
\code{macpan2} that produced \code{model}.
}
\description{
Version of \code{macpan2}
}
