% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_utils.R
\name{mp_zero_vector}
\alias{mp_zero_vector}
\title{Zero Vector}
\usage{
mp_zero_vector(x, ...)
}
\arguments{
\item{x}{Object representing the names of the output vector. Most
commonly this will be a \code{\link{character}} vector.}
}
\description{
Create a \code{\link{numeric}} vector of all zeros with names
given by \code{x}
}
