% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namer.R
\name{Namer}
\alias{Namer}
\title{Namer}
\usage{
Namer()
}
\value{
An object of class \code{Namer} with two methods public,
\code{names} and \code{partitions}, which are inverses of each other.
}
\description{
Name model variables and parse names into labelled partitions.
}
