% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_function.R
\name{ObjectiveFunction}
\alias{ObjectiveFunction}
\title{Objective Function}
\usage{
ObjectiveFunction(obj_fn_expr)
}
\arguments{
\item{obj_fn_expr}{One sided \code{\link{formula}} giving the objective
function of a TMB model. The right hand side expression must contain only
the names of matrices in the model, functions defined in \code{macpan2.cpp},
and numerical literals (e.g. \code{3.14}).}
}
\value{
Object of class \code{ObjectiveFunction} with the following methods.
\subsection{Methods}{
\itemize{
\item \code{data_arg(..., .existing_literals)} -- Return the following components of the data structure
to pass to C++.
\itemize{
\item \code{o_table_x} -- Objective function parse table column giving an index for looking up either
function, matrix, or literal.
\item \code{o_table_n} -- Objective function parse table column giving the number of arguments in
functions.
\item \code{o_table_i} -- Objective function parse table column giving indices for looking up the
rows in the parse table corresponding with the first argument of the
function.
\item \code{literals} -- Numeric vector of literals that can were used in the
expressions of the model.
}
}
}

\subsection{Method Arguments}{
\itemize{
\item \code{...}: Character vector containing the names of the matrices in the model.
\item \code{.existing_literals}: Numeric vector giving the literals used in the
model expressions produced before the objective function.
}
}
}
\description{
Define the objective function of a compartmental model in TMB.
}
