% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_model.R
\name{OptParamsList}
\alias{OptParamsList}
\title{Optimization Parameters List}
\usage{
OptParamsList(
  ...,
  par_id = integer(0L),
  mat = character(0L),
  row_id = integer(0L),
  col_id = integer(0L)
)
}
\arguments{
\item{...}{Objects that can be coerced to numeric vectors, which will be
concatenated to produce the parameter vector.}

\item{par_id}{Integer vector identifying elements of the parameter vector
to be used to replace elements of the model matrices.}

\item{mat}{Character vector the same length as \code{par_id} giving the names of
the matrices containing the elements to replace.}

\item{row_id}{Integer vector the same length as \code{par_id} giving the row
indices of the matrix elements to replace with parameter values.}

\item{col_id}{Integer vector the same length as \code{par_id} giving the column
indices of the matrix elements to replace with parameter values.}
}
\value{
Object of class \code{OptParamsList} with the following methods.
\subsection{Methods}{
\itemize{
\item \verb{$data_arg(..., .type_string = c("p", "r"))}: Return the following components of the data structure
to pass to C++.
\itemize{
\item \verb{\{.type_string\}_par_id} -- Integers identifying the replacing parameter.
\item \verb{\{.type_string\}_mat_id} -- Integers identifying the matrix within which
an element is to be replaced.
\item \verb{\{.type_string\}_row_id} -- Integers identifying the rows within matrices
to replace.
\item \verb{\{.type_string\}_col_id} -- Integers identifying the columns within
matrices to replace.
}
\item \verb{$vector()}: Return the initial value of the numerical parameter vector.
\item \verb{$data_frame()}: Return a data frame with each row describing a parameter.
}
}

\subsection{Method Arguments}{
\itemize{
\item \code{...}: Character vector containing the names of the matrices in the model.
\item \code{.type_string}: Either \code{"p"} or \code{"r"} indicating whether the object
is to be used to represent fixed parameters to be optimized or random
parameters to be integrated out using the Laplace transform.
}
}
}
\description{
Create an object for specifying matrix elements to be optimized or integrated
out of the objective function using a Laplace transform.
}
