% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbparam.R
\name{params_default}
\alias{params_default}
\alias{params_default<-}
\title{Default Parameter Vector}
\usage{
params_default(model)

params_default(model) <- value
}
\arguments{
\item{model}{\code{\link{flexmodel}} object}

\item{value}{object the same type and format as \code{model$params}
\code{new_params} will be merged with \code{model$params}}
}
\description{
Get and set current default value for the parameter vector of
a \code{\link{flexmodel}}
}
