% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_partitions.R
\name{Partition}
\alias{Partition}
\title{Partition}
\usage{
Partition(frame)
}
\arguments{
\item{frame}{Data frame representing the partition. The column names must
consist only of letters, numbers, and start with a letter. The columns of
the data frame must be character vectors such that each value is composed
entirely of letters, numbers, underscore, and must start with a letter
unless it is a blank string. Missing values are not allowed, but blank
strings are. Each row must be unique.}
}
\value{
Object of class \code{Partition} with the following methods.
\subsection{Methods}{
\itemize{
\item \verb{$frame()} -- The \code{Partition} object as a data frame.
\item \verb{$dotted()} -- The \code{Partition} object as a data frame with one united column.
\item \verb{$names()} -- The names of the \code{Partition} (i.e. the column names).
\item \verb{$name()} -- The name of the \code{Partition} (i.e. the dot-concatenated column names).
\item \verb{$labels()} -- The labels of the \code{Partition} (i.e. the row-wise dot-concatenated columns).
\item \verb{$partial_labels()} -- TODO
\item \verb{$filter(..., .wrt, .comparison_function)} -- Filter by keeping only a subset of labels.
\item \verb{$filter_out(..., .wrt, .comparison_function)} -- Filter by removing a subset of labels.
\item \verb{$filter_ordered(..., .wrt, .comparison_function = all_equal)} -- Filter and order by labels.
\itemize{
\item \code{...} -- Labels to filter
\item \code{.wrt} -- The filtering labels are with respect to a particular \code{Partition}, and
\code{.wrt} is the name of this \code{Partition}.
\item \code{.comparison_function} -- Boolean function to decide if each filtering label is
equal to each label in the \code{Partition}.
}
\item \verb{$select(...)} -- Create a new \code{Partition} with a subset of names. The rows of the new
\code{Partition} are de-duplicated.
\item \verb{$select_out(...)} -- Create a new \code{Partition} without a subset of names.
\itemize{
\item \code{...} -- Names to keep in the resulting \code{Partition}.
}
\item \verb{$expand(name)} -- TODO
\item \verb{$union(other)} -- TODO
}
}

\subsection{Fields}{
\itemize{
\item \code{products}
}
}
}
\description{
Create object for manipulating partitions, which are sets of
labels for representing and naming model entities.
}
