% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{rbf}
\alias{rbf}
\title{Radial Basis Functions}
\usage{
rbf(time_steps, dimension, scale = time_steps/dimension)
}
\arguments{
\item{time_steps}{number of time steps in the model}

\item{dimension}{number of gaussians in the basis}

\item{scale}{width of the gaussians}
}
\description{
Radial Basis Functions
}
