% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_partitions.R
\name{read_partition}
\alias{read_partition}
\title{Read Partition}
\usage{
read_partition(...)
}
\arguments{
\item{...}{File path components to pass to \code{\link{CSVReader}}, and
subsequently to \code{\link{file.path}}.}
}
\description{
Read a CSV file in as a \code{\link{Partition}}.
}
