% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data_structure.R
\name{Reader}
\alias{Reader}
\alias{CSVReader}
\alias{JSONReader}
\alias{TXTReader}
\title{Reader}
\usage{
Reader(...)

CSVReader(...)

JSONReader(...)

TXTReader(...)
}
\arguments{
\item{...}{Character vectors giving path components.}
}
\description{
Construct objects for reading data.
}
\section{Functions}{
\itemize{
\item \code{CSVReader()}: Read CSV files.

\item \code{JSONReader()}: Read JSON files.

\item \code{TXTReader()}: Read TXT files.

}}
