% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{reader_spec}
\alias{reader_spec}
\title{Reader Spec}
\usage{
reader_spec(file, reader)
}
\arguments{
\item{file}{Path to a single file.}

\item{reader}{One of the functions documented in \code{\link{Reader}}s.}
}
\value{
A function that takes a path in which to file \code{file} and
returns a named list with one element giving an instantiated
\code{\link{Reader}} object with name given by the file name with the
extension removed. The path may optionally be broken into path components,
which are assembled using \code{\link{file.path}}.
}
\description{
Bundle a file path with a \code{\link{Reader}}.
}
\details{
Typically this function is used when defining an object that inherits
from \code{\link{Files}}.
}
