% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data_structure.R
\name{Scalar2Vector}
\alias{Scalar2Vector}
\title{Scalar2Vector}
\usage{
Scalar2Vector(model)
}
\arguments{
\item{model}{Object of type \code{\link{Model}}
\subsection{Methods}{
\itemize{
\item \verb{$vectorizer(expanded_derivation)}
\item \verb{$vectorize()}
}
}}
}
\description{
Construct an object for replacing scalar names within a \code{\link{Model}}
model, with the equivalent vector name.
}
