% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_partitions.R
\name{Select}
\alias{Select}
\title{mp_select = function(core, dimension_name, labelling_names) {
Select(core, dimension_name, labelling_names)
}}
\usage{
Select(subset, dimension_name, labelling_names)
}
\description{
mp_select = function(core, dimension_name, labelling_names) {
Select(core, dimension_name, labelling_names)
}
}
