% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{self_named_list}
\alias{self_named_list}
\title{Self-Named List}
\usage{
self_named_list(...)
}
\arguments{
\item{...}{Character vectors}
}
\description{
Construct a list of strings with names given by those same strings.
}
