% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model_index.R
\name{show_models}
\alias{show_models}
\title{Print a table of contents of available models}
\usage{
show_models(
  dir = system.file("starter_models", package = "macpan2"),
  show_missing = FALSE,
  for_markdown = FALSE
)
}
\arguments{
\item{dir}{directory to list}

\item{show_missing}{(logical) include entries for models with no README information?}

\item{for_markdown}{(logical) format for rendering the table with markdown-formatted
links to model readme files?}
}
\value{
a data frame containing entries \code{dir} (model directory), \code{title} (model title), \code{index_entry} (short description)
}
\description{
Collects information from the headers of the README files in the model directories
and returns the results as a data frame
}
\examples{
show_models(show_missing = TRUE)
}
