% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_caching.R
\name{si_example}
\alias{si_example}
\alias{si_example_object}
\alias{si_example_code}
\title{SI Example}
\usage{
si_example_object(
  object = c("specification", "simulator", "data", "calibrator", "optimized_calibrator")
)

si_example_code(
  object = c("specification", "simulator", "data", "calibrator", "optimized_calibrator")
)
}
\arguments{
\item{object}{Type of object associated with the example SI model. Can be
one of the following:
\code{"specification"}, \code{"simulator"}, \code{"data"}, \code{"calibrator"},
\code{"optimized_calibrator"}.
The functions used to produce each of these objects, respectively, are
\code{\link{mp_tmb_library}},
\code{\link{mp_simulator}},
\code{\link{mp_trajectory}},
\code{\link{mp_tmb_calibrator}}, and \code{\link{mp_tmb_calibrator}}.}
}
\description{
The \code{si_example_object} function efficiently generates objects
associated with the
\href{https://github.com/canmod/macpan2/blob/main/inst/starter_models/si/README.md}{example SI model}.
It either creates these objects or extracts them from a pre-computed cache if
they have already been created in the current session. This function is
predominantly used to simplify examples in the package documentation -- it
allows package developers to efficiently get these basic example objects to
illustrate many different concepts, without having to spend time computing
them over and over again when generating the documentation. The
\code{si_example_code} function displays code
that could be used to generate these objects, and clarifies what these
functions do.
}
\section{Functions}{
\itemize{
\item \code{si_example_object()}: Return an object associated with the example
SI model.

\item \code{si_example_code()}: Show code that could be used to generate an object
associated with the example SI model.

}}
\examples{

# Get SI calibrator in two different states
cal1 = si_example_object("calibrator")
cal2 = si_example_object("optimized_calibrator")

# Optimize cal1 so that it is equal to cal2
mp_optimize(cal1)
mp_tmb_coef(cal1)
mp_tmb_coef(cal2)

# Inspect the code used to produce all objects in this example SI model.
si_example_code("specification")
si_example_code("simulator")
si_example_code("data")
si_example_code("calibrator")
si_example_code("optimized_calibrator")

}
