% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_list_generators.R
\name{SimpleChangeModel}
\alias{SimpleChangeModel}
\title{Simple Change Model}
\usage{
SimpleChangeModel(before = list(), during = list(), after = list())
}
\arguments{
\item{before}{List of two-sided formulas to be run in the before phase.}

\item{during}{List of two-sided formulas and/or \code{ChangeComponent} objects
to be used to create a \code{during} expression list after choosing an
\code{UpdateMethod}.}

\item{after}{List of two-sided formulas to be run in the after phase.}
}
\description{
Simple Change Model
}
\examples{

si = SimpleChangeModel(
  during = list(mp_per_capita_flow("S", "I", infection ~ beta * I / N))
) |> EulerUpdateMethod()
si$during()

}
