% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_model_constructors.R
\name{SimulatorConstructor}
\alias{SimulatorConstructor}
\title{Simulator Constructor}
\usage{
SimulatorConstructor(model_directory, integration_method = RK4, ...)
}
\arguments{
\item{model_directory}{A string giving a path to a directory containing
model definition files.}

\item{integration_method}{One of the functions described in
\link{integration_methods}, used to integrate the dynamical system.}

\item{...}{Arguments to pass to \code{\link{TMBModel}}.}
}
\description{
Simulator Constructor
}
