% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulators.R
\name{Simulators}
\alias{Simulators}
\title{Simulators}
\usage{
Simulators(model)
}
\arguments{
\item{model}{A \code{\link{Compartmental}} model.}
}
\value{
An object with the following methods.
\subsection{Methods}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{$tmb(time_steps
   , state, rate
   , ...
   , .mats_to_save = .mats_to_return
   , .mats_to_return = "state"
   , .dimnames = list()
)
}\if{html}{\out{</div>}}
}

\subsection{Method Arguments}{
\itemize{
\item \code{time_steps}: The number of time steps run in each simulation.
\item \code{state}: Named numerical vector giving the default initial value of the
state vector. The names must be equal to \code{model$state_labels()}.
\item \code{flow}: Named numerical vector giving the default initial value of the
flow vector. The names must be equal to \code{model$flow_labels()}.
\item \code{...}: Named numerical objects that can be coerced to a matrix, giving
the matrices required by the model.
\item \code{.mats_to_save}: Character vector of names of matrices to save. Defaults
to \code{"state"}, which is the state vector. Other useful names include
\code{"total_inflow"} (the incidence associated with each state variable)
and \code{"total_outflow"} (the total leaving each state variable at each
time step). One may also add any variable in \code{model$other_labels()}.
See \code{\link{MatsList}} for details.
\item \code{.mats_to_return}: Character vector of names of matrices to return. Defaults
to \code{"state"}, which is the state vector. Other useful names include
\code{"total_inflow"} (the incidence associated with each state variable)
and \code{"total_outflow"} (the total leaving each state variable at each
time step). One may also add any variable in \code{model$other_labels()}.
See \code{\link{MatsList}} for details.
\item \code{.dimnames}: Advanced and often not necessary.  See
\code{\link{MatsList}} for details.
}
}
}
\description{
Adapt a \code{\link{Compartmental}} model to a simulation engine. Currently
TMB is the only simulation engine.
}
