% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data_structure.R
\name{StandardExpr}
\alias{StandardExpr}
\title{StandardExpr}
\usage{
StandardExpr(model)
}
\arguments{
\item{model}{Object created by \code{\link{Model}}.}
}
\description{
Evaluate standard model expressions
}
