% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_to_models.R
\name{StandardExprHazard}
\alias{StandardExprHazard}
\title{Standard Expressions for Hazard-Corrected Models}
\usage{
StandardExprHazard(model)
}
\arguments{
\item{model}{Object created by \code{\link{Model}}.}
}
\description{
Evaluate standard expressions in hazard-corrected models.
}
