% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_and_labels.R
\name{StringData}
\alias{StringData}
\alias{StringDataFromFrame}
\alias{StringDataFromDotted}
\alias{print.StringData}
\title{String Data}
\usage{
StringDataFromFrame(data)

StringDataFromDotted(labels, name)

\method{print}{StringData}(x, ...)
}
\arguments{
\item{data}{Data frame with names given by column names and labels by
the elements of the columns.}

\item{labels}{Character vector with (dot-separated) partition labels.}

\item{name}{Character scalar with (dot-separated) partition name.}

\item{x}{\code{StringData} object}

\item{...}{Not used but present for S3 method consistency.}
}
\description{
Create objects for representing names and labels in a dynamical
model.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print(StringData)}: Print out a \code{StringData} object.

}}
\section{Functions}{
\itemize{
\item \code{StringDataFromFrame()}: Construct object from a data frame without any
dots in either the names or the values.

\item \code{StringDataFromDotted()}: Construct object from a character scalar with
(dot-separated) partition names and a character vector with (dot-separated)
partition labels.

}}
\examples{
vars = (mp_cartesian(
      mp_index(Epi = c("S", "I", "R"))
    , mp_index(Age = c("young", "old"))
  )
  |> as.data.frame()
  |> StringDataFromFrame()
)
vars
vars$dot()

}
