% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolizer.R
\name{symbolizer}
\alias{symbolizer}
\title{Symbolizer}
\usage{
symbolizer(f)
}
\arguments{
\item{f}{function that takes numeric arguments and returns
a numeric output}
}
\description{
Convert a numeric function into one that can symbolically
manipulate strings as though they were mathematical symbols.
}
\details{
Only certain functions are allowed to be utilized in the
mathematical function. See the examples for how to get
a list of these functions.
}
\examples{
gg = function(S, I, R, beta) beta * S * I / (S + I + R)
gg(1, 0.1, 0, 0.1)
symbolizer(gg)(1, 0.1, 0, 0.1)
ff = function(x, i, j) x[i, j] / sum(x)
ff(matrix(1, 3, 3), 2, 2)
symbolizer(ff)("x", 2, 2)
gs = symbolizer(gg)
adder = symbolizer(function(x, y) (x) + (y))
adder(
  gs("S_unvax", "I_unvax", "R_unvax", "beta_unvax"),
  gs("S_unvax", "I_vax", "R_vax", "beta_vax")
)
ff = symbolizer(function(x, y, z) {
  A = x + y
  B = z^2
  A / B
})
ff(1, 2, 3)

# allowed functions:
ls(environment(ff))
}
