% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{Time}
\alias{Time}
\title{Time}
\usage{
Time(time_steps)
}
\arguments{
\item{time_steps}{Number of time steps in the simulation loop.}
}
\value{
Object of class \code{Time} with the following methods.
\subsection{Methods}{
\itemize{
\item \verb{$data_arg()} -- Return the following components of the data structure
to pass to C++.
\itemize{
\item \code{time_steps} -- Number of time steps in the simulation loop.
}
}
}
}
\description{
Define the number of time steps in a compartmental model in TMB.
}
