% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_model.R
\name{TMBSimulator}
\alias{TMBSimulator}
\title{TMB Simulator}
\usage{
TMBSimulator(tmb_model, tmb_cpp = "macpan2")
}
\arguments{
\item{tmb_model}{An object of class \code{\link{TMBModel}}.}

\item{tmb_cpp}{Name of a C++ program using TMB as the simulation engine.}
}
\value{
Object of class \code{TMBSimulator} with the following methods.
\subsection{Methods}{
\itemize{
\item \verb{$ad_fun()} -- TODO
\item \verb{$report()} -- TODO
}
}
}
\description{
Construct an object with methods fore simulating from and optimizing a
compartmental model made using \code{\link{TMBModel}}.
}
