% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_tmb_calibrator.R
\name{TMBTV.RBFArg}
\alias{TMBTV.RBFArg}
\title{TMBTV.TVArg = function(
tv
, struc
, spec
, existing_global_names = character()
) {
self = TMBTVAbstract()
self$existing_global_names = existing_global_names
self$spec = spec

self$before_loop = function() list()
self$after_loop = function() list()
\subsection{List with the values of each}{
}

\subsection{time varying parameter at the change points. The}{
}

\subsection{names of the list are the time-varying matrices}{
}

\subsection{in the spec.}{

self$time_var = function() list()
}

\subsection{List of the integers}{
}

\subsection{giving the time-steps of the changepoints with}{
}

\subsection{the first time-step always being 0 (the initial)}{
}

\subsection{The names of the list are the time-varying}{
}

\subsection{matrices in the spec.}{

self$change_points = function() list()
self$change_pointer = function() list()
}

\subsection{List of expressions that update parameters that}{
}

\subsection{are time-varying}{

self$var_update_exprs = function() list()
}

\subsection{data frames describing the fixed and random effects corresponding}{
}

\subsection{to time-varying parameters}{

self$tv_params_frame = function(tv_pars) self$empty_params_frame
self$tv_random_frame = function() self$empty_params_frame

return_object(self, "TMBTV")
}
}}
\usage{
\method{TMBTV}{RBFArg}(tv, struc, spec, existing_global_names = character())
}
\description{
TMBTV.TVArg = function(
tv
, struc
, spec
, existing_global_names = character()
) {
self = TMBTVAbstract()
self$existing_global_names = existing_global_names
self$spec = spec

self$before_loop = function() list()
self$after_loop = function() list()
\subsection{List with the values of each}{
}

\subsection{time varying parameter at the change points. The}{
}

\subsection{names of the list are the time-varying matrices}{
}

\subsection{in the spec.}{

self$time_var = function() list()
}

\subsection{List of the integers}{
}

\subsection{giving the time-steps of the changepoints with}{
}

\subsection{the first time-step always being 0 (the initial)}{
}

\subsection{The names of the list are the time-varying}{
}

\subsection{matrices in the spec.}{

self$change_points = function() list()
self$change_pointer = function() list()
}

\subsection{List of expressions that update parameters that}{
}

\subsection{are time-varying}{

self$var_update_exprs = function() list()
}

\subsection{data frames describing the fixed and random effects corresponding}{
}

\subsection{to time-varying parameters}{

self$tv_params_frame = function(tv_pars) self$empty_params_frame
self$tv_random_frame = function() self$empty_params_frame

return_object(self, "TMBTV")
}
}
}
