% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_utils.R
\name{to_labels}
\alias{to_labels}
\title{To Labels}
\usage{
to_labels(x)
}
\arguments{
\item{x}{Object to convert to labels.}
}
\value{
Character vector that can be used as labels.
}
\description{
Convert objects to labels, which are vectors that can be dotted.
}
