% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_utils.R
\name{to_name}
\alias{to_name}
\title{To Name}
\usage{
to_name(x)
}
\arguments{
\item{x}{Object to convert to labels.}
}
\value{
Character string that can be used as a name.
}
\description{
Convert objects to a name, which is a scalar string that can be dotted.
}
