% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_utils.R
\name{to_names}
\alias{to_names}
\title{To Names}
\usage{
to_names(x)
}
\arguments{
\item{x}{Object to convert to names.}
}
\value{
Character vector that can be used as names.
}
\description{
Convert objects to names, which are character vectors with the following
restrictions:  (1) they cannot have dots, (2) all values must start with
a letter, (3) all characters must be letters, numbers, or underscore.
}
