% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_utils.R
\name{to_positions}
\alias{to_positions}
\title{To Positions}
\usage{
to_positions(x)
}
\arguments{
\item{x}{An object of a class that can be converted to a
position vector.}
}
\description{
Return position vector of indices corresponding to the
input object.
}
\seealso{
\code{\link[=mp_positions]{mp_positions()}}
}
