% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_utils.R
\name{to_string}
\alias{to_string}
\title{To String}
\usage{
to_string(x)
}
\arguments{
\item{x}{Object to convert to a string.}
}
\value{
A length-1 character vector.
}
\description{
Convert an object to a string.
}
