% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\docType{data}
\name{transform_distr_param}
\alias{transform_distr_param}
\alias{mp_identity}
\alias{mp_log}
\title{Distributional Parameter Transformation}
\usage{
mp_identity

mp_log
}
\description{
\itemize{
\item \code{mp_identity} - Identity transformation
}

\itemize{
\item \code{mp_log} - Log transformation
}
}
\keyword{datasets}
