% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_partitions.R
\name{union_vars}
\alias{union_vars}
\title{Union of Variables}
\usage{
union_vars(...)
}
\arguments{
\item{...}{\code{\link{Partition}} objects to combine.}
}
\description{
Take the union of a set of variable lists, each of which is represented
by a \code{\link{Partition}} object.
}
