% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_list_generators.R
\name{update_methods}
\alias{update_methods}
\alias{EulerUpdateMethod}
\alias{RK4UpdateMethod}
\alias{EulerMultinomialUpdateMethod}
\title{Update Methods}
\usage{
EulerUpdateMethod(change_model, existing_global_names = character())

RK4UpdateMethod(change_model)

EulerMultinomialUpdateMethod(change_model)
}
\arguments{
\item{change_model}{An object of class \code{ChangeModel}.}
}
\description{
Convert a \code{ChangeModel} into an object that can return \code{before}, \code{during},
and \code{after} expression lists.
}
\section{Functions}{
\itemize{
\item \code{RK4UpdateMethod()}: Create expression lists for solving ODEs using
Runge Kutta 4.

\item \code{EulerMultinomialUpdateMethod()}: Create expression lists for generating process
error using the Euler Multinomial distribution.

}}
