% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/products.R
\name{cartesian}
\alias{cartesian}
\title{Cartesian Product}
\usage{
cartesian(x, y)
}
\arguments{
\item{x}{A \code{\link{Partition}} object.}

\item{y}{A \code{\link{Partition}} object.}
}
\description{
Cartesian Product
}
