% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compartmental.R
\name{Compartmental}
\alias{Compartmental}
\title{Compartmental Model}
\usage{
Compartmental(model_directory)
}
\arguments{
\item{model_directory}{String giving a path to a directory containing
the following files, \code{variables.csv}, \code{derivations.json}, \code{flows.csv},
and \code{settings.json}, described by
\href{https://canmod.github.io/macpan2/articles/model_definitions.html}{this spec}.}
}
\description{
Create an object for containing a compartmental model.
}
