% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data_structure.R
\name{DerivationExpander}
\alias{DerivationExpander}
\title{DerivationExpander}
\usage{
DerivationExpander(model)
}
\arguments{
\item{model}{Object of type \code{\link{Model}}}
}
\description{
Construct an object for expanding the derivations within a
\code{\link{Model}}.
}
