% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_to_models.R
\name{Derivations2ExprList}
\alias{Derivations2ExprList}
\title{Convert Derivation Lists to Expression Lists}
\usage{
Derivations2ExprList(user_expr, standard_expr)
}
\arguments{
\item{user_expr}{\code{\link{UserExpr}} object.}

\item{standard_expr}{\code{\link{StandardExpr}} object.}
}
\value{
Object of class \code{Derivations2ExprList} with the following
methods.
\subsection{Methods}{
\itemize{
\item \verb{$expr_list()} -- An alternate constructor of \code{\link{ExprList}}
objects from a set of derivations.
}
}

\subsection{Arguments}{
\itemize{
\item \code{.simulate_exprs} -- See the argument of the same name in
\code{\link{ExprList}}.
}
}
}
\description{
Convert Derivation Lists to Expression Lists
}
