% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{distribution}
\alias{distribution}
\alias{mp_uniform}
\alias{mp_normal}
\alias{mp_log_normal}
\alias{mp_poisson}
\alias{mp_neg_bin}
\title{Distributions}
\usage{
mp_uniform()

mp_normal(location = DistrParam("location"), sd)

mp_log_normal(location = DistrParam("location"), sd)

mp_poisson(location = DistrParam("location"))

mp_neg_bin(location = DistrParam("location"), disp)
}
\arguments{
\item{location}{Location parameter. Only necessary if used as a prior
distribution. If it is used as a likelihood component the location
parameter will be taken as the simulated variable being fitted to data,
and so this \code{location} parameter should be left to the default.}

\item{sd}{Standard deviation parameter.}

\item{disp}{Dispersion parameter.}
}
\description{
Distributions which can be used to specify prior or likelihood components in
model calibration.

\itemize{
\item Uniform Distribution (Improper), only appropriate for prior components - \code{mp_uniform}
}

\itemize{
\item Normal Distribution - \code{mp_normal}
}

\itemize{
\item Log-Normal Distribution - \code{mp_log_normal}
}

\itemize{
\item Poisson Distribution - \code{mp_poisson}
}

\itemize{
\item Negative Binomial Distribution - \code{mp_neg_bin}
}
}
