% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{distribution}
\alias{distribution}
\alias{mp_uniform}
\alias{mp_normal}
\alias{mp_log_normal}
\alias{mp_logit_normal}
\alias{mp_poisson}
\alias{mp_neg_bin}
\title{Distributions}
\usage{
mp_uniform(trans_distr_param = list())

mp_normal(
  location = mp_distr_param_null("location"),
  sd,
  trans_distr_param = list(location = mp_identity, sd = mp_log)
)

mp_log_normal(
  location = mp_distr_param_null("location"),
  sd,
  trans_distr_param = list(location = mp_identity, sd = mp_identity)
)

mp_logit_normal(
  location = mp_distr_param_null("location"),
  sd,
  trans_distr_param = list(location = mp_identity, sd = mp_identity)
)

mp_poisson(
  location = mp_distr_param_null("location"),
  trans_distr_param = list(location = mp_identity)
)

mp_neg_bin(
  location = mp_distr_param_null("location"),
  disp,
  trans_distr_param = list(location = mp_identity, disp = mp_log)
)
}
\arguments{
\item{trans_distr_param}{Named list of transformations for each
distributional parameter. See \code{\link{transform_distr_param}} for a
list of available transformations.}

\item{location}{Location parameter.
Specifying the \code{location} parameter is only necessary when the distribution
is used as a prior distribution. If it is used as a likelihood component the
location parameter will be taken as the simulated variable being fitted to
data, and so this \code{location} parameter should be left to the default.}

\item{sd}{Standard deviation parameter.}

\item{disp}{Dispersion parameter.}
}
\description{
Distributions which can be used to specify prior or likelihood components in
model calibration.

\itemize{
\item Uniform Distribution (Improper), only appropriate for prior
components - \code{mp_uniform}
}

\itemize{
\item Normal Distribution - \code{mp_normal}
}

\itemize{
\item Log-Normal Distribution - \code{mp_log_normal}
}

\itemize{
\item Logit-Normal Distribution - \code{mp_logit_normal}
}

\itemize{
\item Poisson Distribution - \code{mp_poisson}
}

\itemize{
\item Negative Binomial Distribution - \code{mp_neg_bin}
}
}
\details{
All distributional parameter arguments can be specified either as
a numeric value, a character string giving the parameter name, or a
distributional parameter object (See ?fit_distr_params).
}
