% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{DistrSpec}
\alias{DistrSpec}
\title{Distribution Specification}
\usage{
DistrSpec(distr_param_objs = list(), default_trans = list())
}
\arguments{
\item{distr_param_objs}{List of DistrParam objects defining the distribution. All
distributions must have a \code{location} parameter as the first parameter when
the distribution is instantiated. However, this \code{location} parameter
might be removed by the \code{remove_location_parameter} method, which
should be called later if it is determined that the distribution does
not require a \code{location} parameter (e.g., it is a likelihood component
that takes the simulated trajectory as the location).
data frame characterizing what distributional parameters associated
with this distribution should be optimized as fixed effects.
data frame characterizing what distributional parameters associated
with this distribution should be optimized as random effects.
character string giving an expression for the component of a log prior
density associated with this distribution. this may not be used if
it is intended to be used as a likelihood component.
character string giving an expression for the component of a log likelihood
function associated with this distribution. this may not be used if
it is intended to be used as a prior distribution component.
for the future (e.g. beta ~ rnorm(0, 1))}
}
\description{
Class representing the specification of a distribution, either a prior or
likelihood component.
}
\details{
Extend this class to develop new distributions.
}
