% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout-plot-utils.R
\name{dot_layout}
\alias{dot_layout}
\title{Create a graph from a model specification
Converts a model specification into a graph (using the \code{graph} package) that can be plotted with \code{Rgraphviz}: see \code{?Rgraphviz::plot.graphNEL} and https://graphviz.org/doc/info/attrs.html for information on customizing the plot}
\usage{
dot_layout(spec, ...)
}
\arguments{
\item{spec}{a model specification}
}
\description{
Create a graph from a model specification
Converts a model specification into a graph (using the \code{graph} package) that can be plotted with \code{Rgraphviz}: see \code{?Rgraphviz::plot.graphNEL} and https://graphviz.org/doc/info/attrs.html for information on customizing the plot
}
\examples{
if (require(Rgraphviz)) {
  macpan_base = mp_tmb_library("starter_models", "macpan_base", package = "macpan2")
  dot_layout(macpan_base) |>
    plot(attrs = list(graph = list(rankdir = "LR"),
                      node = list(shape = "rectangle")))
}
}
