% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_expr.R
\docType{data}
\name{empty_matrix}
\alias{empty_matrix}
\title{Empty Matrix}
\format{
A \code{\link{numeric}} \code{\link[base]{matrix}} with zero rows
and zero columns.
}
\usage{
empty_matrix
}
\description{
Empty matrices are useful when defining matrices that do not need to be
initialized because they will get computed before they are required by
other expressions. They are also a useful placeholder for matrices that
should only have a value after a certain phase in the simulation.
}
\examples{
s = TMBModel(
  init_mats = MatsList(x = empty_matrix
    , .mats_to_save = "x"
    , .mats_to_return = "x"
  ),
  expr_list = ExprList(during = list(x ~ time_step(0))),
  time_steps = Time(2)
)
s$simulator()$report()
}
\keyword{datasets}
