% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_eval.R
\name{engine_eval}
\alias{engine_eval}
\title{Engine Evaluation}
\usage{
engine_eval(
  e,
  ...,
  .matrix_to_return,
  .tmb_cpp = "macpan2",
  .structure_labels = NullLabels()
)
}
\arguments{
\item{e}{Expression as a one-sided formula, the right-hand-side of which is
treated as the expression to be evaluated.}

\item{...}{Named objects that can be coerced into numeric matrices.}

\item{.matrix_to_return}{Optional name of one of the matrices given
in \code{...} to be returned. If this argument is missing, the
matrix that will be returned is the matrix returned by the expression on
the right-hand-side of the formula.}

\item{.tmb_cpp}{Name of a \code{C++} program defining the engine. Typically
you just want to use the default, which is \code{macpan2}, unless you
are extending the
\href{https://canmod.github.io/macpan2/articles/cpp_side.html}{engine}
yourself.}

\item{.structure_labels}{See \code{\link{MatsList}}.}
}
\value{
Matrix being produced on the right-hand-side or matrix given in
\code{.matrix_to_return} if it is provided.
}
\description{
Evaluate an expression in the TMB-based C++ simulation and objective
function engine. This function is useful for trying out the
\code{\link{engine_functions}} that can be used to define \pkg{macpan2}
models.
}
\examples{
engine_eval(~ exp(y), y = pi) # ~ 23.14069

# It is not currently possible to assign values to a subset of
# a matrix in a natural way (e.g. you cannot do things like x[1] = exp(y)),
# but you can achieve this functionality using the assign function.
engine_eval(~ assign(x, 1, 0, exp(y))
  , x = rep(0, 2)
  , y = pi
  , .matrix_to_return = "x"
)
}
