% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macpan2.R
\docType{package}
\name{macpan2-package}
\alias{macpan2}
\alias{macpan2-package}
\title{\pkg{macpan2}}
\description{
This R package is a do-over of the
\href{https://github.com/mac-theobio/mcmasterpandemic}{McMasterPandemic}
package on compartmental epidemic modelling for
forecasting and analysis of infectious diseases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://canmod.github.io/macpan2/}
  \item \url{https://github.com/canmod/macpan2}
  \item Report bugs at \url{https://github.com/canmod/macpan2/issues}
}

}
\author{
\strong{Maintainer}: Steve Walker \email{swalk@mcmaster.ca}

Authors:
\itemize{
  \item Weiguang Guan
  \item Ben Bolker
  \item Jen Freeman
  \item Darren Flynn-Primrose
}

Other contributors:
\itemize{
  \item Irena Papst [contributor]
  \item Michael Li [contributor]
}

}
\keyword{internal}
