% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_expr.R
\name{make_expr_parser}
\alias{make_expr_parser}
\title{Generate an Arithmetic Expression Parser}
\usage{
make_expr_parser(parser_name = "parse_expr", finalizer = force)
}
\arguments{
\item{valid_funcs}{character vector with names of valid functions}

\item{valid_vars}{character vector with names of valid variables}
}
\value{
A function that accepts a call or list of expressions, and returns
a data frame with rows representing the nodes of the parse tree
and the following columns \describe{
\item{x}{name of the function, variable, or literal}
\item{n}{number of arguments if \code{x} is a function, zero otherwise}
\item{i}{row index giving the node corresponding to the first argument
of the function -- zero if \code{x} is not a function}
\item{q}{boolean indicating if the node corresponds to a literal}
}
}
\description{
Generate an Arithmetic Expression Parser
}
