% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_files.R
\name{ModelCollection}
\alias{ModelCollection}
\title{Model Collection}
\usage{
ModelCollection(variables, derivations, flows, settings)
}
\arguments{
\item{variables}{Return value of the \code{variables} method in a
\code{\link{ModelFiles}} object.}

\item{derivations}{Return value of the \code{derivations} method in a
\code{\link{ModelFiles}} object.}

\item{flows}{Return value of the \code{flows} method in a
\code{\link{ModelFiles}} object.}

\item{settings}{Return value of the \code{settings} method in a
\code{\link{ModelFiles}} object.}
}
\value{
An object of class \code{ModelCollection} with the following
methods.
\subsection{Methods}{
\itemize{
\item \verb{$variables()} --
\item \verb{$derivations()}
\item \verb{$flows()}
\item \verb{$settings()}
}
}
}
\description{
A model definition that is untied from a set of \code{\link{ModelFiles}}.
}
