% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_group.R
\name{mp_aggregate}
\alias{mp_aggregate}
\title{Aggregate an Index}
\usage{
mp_aggregate(index, by = "Group", ledger_column = "group")
}
\description{
Create a one-column ledger (see \code{\link{LedgerDefinition}}) with rows
identifying instances of an aggregation.
}
\seealso{
Other functions that return ledgers
\code{\link{mp_join}()}
}
\concept{ledgers}
