% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{mp_aggregate}
\alias{mp_aggregate}
\title{Aggregate an Index}
\usage{
mp_aggregate(index, by = "Group", ledger_column = "group")
}
\description{
Create
}
\seealso{
Other functions that return ledgers
\code{\link{LedgerDefinition}},
\code{\link{mp_join}()}
}
\concept{ledgers}
