% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_cartesian.R
\name{mp_cartesian}
\alias{mp_cartesian}
\title{Cartesian Product of Index Tables}
\usage{
mp_cartesian(...)
}
\arguments{
\item{...}{Index tables (see \code{\link{mp_index}}).}
}
\description{
Produce a new index table by taking all possible pairwise combinations
of the input tables. This is useful for producing product models
that expand model components through stratification.
}
\examples{
mp_cartesian(
  mp_index(Epi = c("S", "I")),
  mp_index(Age = c("young", "old"))
)

si = mp_index(Epi = c("S", "I"))
age = mp_index(Age = c("young", "old"))
loc = mp_index(City = c("hamilton", "toronto"))
vax = mp_index(Vax = c("unvax", "vax"))
(si
  |> mp_cartesian(age)
  |> mp_cartesian(loc)
  |> mp_cartesian(vax)
)

flow_rates = mp_index(Epi = c("infection", "recovery"))
mp_union(
  mp_cartesian(
    mp_subset(flow_rates, Epi = "infection"),
    age
  ),
  mp_subset(flow_rates, Epi = "recovery")
)

}
\seealso{
Other functions that return index tables
\code{\link{mp_index}()},
\code{\link{mp_rename}()},
\code{\link{mp_subset}()},
\code{\link{mp_union}()}

Other functions that take products of index tables and return one index tables
\code{\link{mp_linear}()},
\code{\link{mp_square}()},
\code{\link{mp_symmetric}()},
\code{\link{mp_triangle}()}
}
\concept{indexes}
\concept{products}
