% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_model.R
\name{mp_default}
\alias{mp_default}
\alias{mp_default_list}
\title{Default Values}
\usage{
mp_default(model)

mp_default_list(model)
}
\arguments{
\item{model}{A model object from which to extract default values.}
}
\value{
A long-format data frame with default values for matrices required
as input to model objects. The columns of this output are \code{matrix}, \code{row},
\code{col}, and \code{value}. Scalar matrices do not have any entries in the \code{row} or
\code{col} columns.
}
\description{
Default Values
}
\section{Functions}{
\itemize{
\item \code{mp_default_list()}: List of the default variables as matrices.

}}
