% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_def_run.R
\name{mp_dynamic_model}
\alias{mp_dynamic_model}
\title{Dynamic Model}
\usage{
mp_dynamic_model(
  expr_list = ExprList(),
  ledgers = list(),
  init_vecs = list(),
  unstruc_mats = list()
)
}
\description{
Dynamic Model
}
