% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_to_tmb.R
\name{mp_dynamic_simulator}
\alias{mp_dynamic_simulator}
\title{TMB Simulator from Dynamic Model}
\usage{
mp_dynamic_simulator(
  dynamic_model,
  time_steps = 0L,
  vectors = NULL,
  unstruc_mats = NULL,
  mats_to_save = NULL,
  mats_to_return = NULL,
  params = OptParamsList(0),
  random = OptParamsList(),
  obj_fn = ObjectiveFunction(~0),
  log_file = LogFile(),
  do_pred_sdreport = TRUE,
  tmb_cpp = "macpan2",
  initialize_ad_fun = TRUE,
  ...
)
}
\arguments{
\item{dynamic_model}{Object product by \code{\link{mp_dynamic_model}}.}

\item{time_steps}{Number of time steps to simulate.}

\item{vectors}{Named list of named vectors as initial values for the
simulations that are referenced in the expression list in the dynamic model.}

\item{unstruc_mats}{= Named list of objects that can be coerced to
numerical matrices that are used in the expression list of the
dynamic model.}

\item{mats_to_save}{TODO}

\item{mats_to_return}{TODO}

\item{params}{TODO}

\item{random}{TODO}

\item{obj_fn}{TODO}

\item{log_file}{TODO}

\item{do_pred_sdreport}{TODO}

\item{tmb_cpp}{TODO}

\item{initialize_ad_fun}{TODO}

\item{...}{TODO}
}
\description{
This is an 'old' function that was tested out at a workshop.
Currently it still drives the engine-agnostic-grammar vignette,
but we plan to replace this function with \code{\link{mp_simulator}}.
}
