% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_methods.R
\name{mp_effects_descr}
\alias{mp_effects_descr}
\title{Describe Statistical Effects}
\usage{
mp_effects_descr(model)
}
\arguments{
\item{model}{A model in the TMB engine that can be used to compute
tables of statistical effects.}
}
\description{
Additional information that can be joined to the output of
the \code{tidy.TMB} or \code{tidy.stanfit} functions in the \code{broom.mixed} package.
}
