% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_frame.R
\name{mp_flow_frame}
\alias{mp_flow_frame}
\title{Data Frame Describing Compartmental Model Flows}
\usage{
mp_flow_frame(spec, topological_sort = TRUE, loops = "^$")
}
\arguments{
\item{spec}{A \code{\link{mp_tmb_model_spec}}.}

\item{topological_sort}{Should the states be topologically sorted to
respect the main direction of flow?}

\item{loops}{Pattern for matching the names of flows that make
the flow model not a DAG, which is a critical assumption when topologically
sorting the order of states and flows in the output. This is only relevant if
\code{topological_sort} is used.}
}
\value{
A data frame that gives information provided in calls to
\code{\link{mp_per_capita_flow}} and \code{\link{mp_per_capita_inflow}}.
}
\description{
Get a data frame where each row represents a flow in a model specification.
}
