% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{mp_index}
\alias{mp_index}
\title{Create Index Object}
\usage{
mp_index(..., labelling_names)
}
\arguments{
\item{...}{Character vectors to combine into an \code{\link{Index}}.
object.}

\item{labelling_names}{See \code{\link{Index}}.}
}
\description{
Create an \code{\link{Index}} object from a set of
character vectors. This function is analogous to the
\code{\link{data.frame}} function for creating data frames.
}
\examples{
mp_index(
  Epi = c("S", "I", "S", "I"),
  Age = c("young", "young", "old", "old")
)

}
\seealso{
\code{\link[=Index]{Index()}}
}
