% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_join.R
\name{mp_join}
\alias{mp_join}
\title{Join Indexes}
\usage{
mp_join(..., by = empty_named_list())
}
\arguments{
\item{...}{Named arguments giving indexes created by
\code{\link{mp_index}} or another function that manipulates indexes.
Each argument will become a position vector used to subset
or expand numeric vectors in archetype formulas.}

\item{by}{What columns to use to join the indexes. See below on
how to specify this argument.}
}
\description{
Join two or more index tables (see \code{\link{mp_index}}) to produce a
ledger (see \code{\link{LedgerDefinition}}).
}
\details{
When two index tables are passed to \code{...}, \code{mp_join} behaves very much like
an ordinary \href{https://en.wikipedia.org/wiki/Join_(SQL)}{inner join}.
When more than two tables are passed to \code{...}, \code{mp_join} iteratively joins
pairs of tables to produce a final ledger. For example, if index tables \code{A}
\code{B}, and \code{C} are passed to \code{mp_join}, an inner join of \code{A} and \code{B} is
performed and the result is joined with \code{C}. In each of these successive
internal joins. The properties of inner
joins ensures that the order of tables does not affect the set of rows in
the final table (SW states without proof!).
}
\examples{

# When two index tables are passed to `...`, the `by` argument is just a
# character vector of column names on which to join (as in standard R functions
# for joining data frames), or the dot-concatenation of these column names.
# For example,
state = mp_index(
  Epi = c("S", "I", "S", "I"),
  Age = c("young", "young", "old", "old")
)
mp_join(
  from = mp_subset(state, Epi = "S"),
  to = mp_subset(state, Epi = "I"),
  by = "Age"
)
# If there are more than two tables then the `by` argument must be a named
# list of character vectors, each describing how to join the columns of
# a pair of tables in `...`. The names of this list are dot-concatenations
# of the names of pairs of tables in `...`. For example,
rates = mp_index(
  Epi = c("lambda", "lambda"),
  Age = c("young", "old")
)
mp_join(
  from = mp_subset(state, Epi = "S"),
  to = mp_subset(state, Epi = "I"),
  rate = mp_subset(rates, Epi = "lambda"),
  by = list(
    from.to = "Age",
    from.rate = "Age"
  )
)
# If the `by` columns have different names in two tables, then you can
# specify these using formula notation where the left-hand-side
# is a dot-concatenation of columns in the first table and the
# right-hand-side is a dot-concatenation of the columns in the second
# table. For example,
contact = mp_index(
  AgeSusceptible = c("young", "young", "old", "old"),
  AgeInfectious = c("young", "old", "young", "old")
)
mp_join(
  sus = mp_subset(state, Epi = "S"),
  inf = mp_subset(state, Epi = "I"),
  con = contact,
  by = list(
    sus.con = "Age" ~ "AgeSusceptible",
    inf.con = "Age" ~ "AgeInfectious"
  )
)
}
\seealso{
Other functions that return ledgers
\code{\link{mp_aggregate}()}
}
\concept{ledgers}
\concept{population-structure}
