% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_operator.R
\name{mp_kronecker_operator}
\alias{mp_kronecker_operator}
\title{Kronecker Operator}
\usage{
mp_kronecker_operator(operator)
}
\arguments{
\item{operator}{A scalar binary operator.}
}
\value{
A Kronecker operator convenient for use with \code{macpan2}.
}
\description{
Convert a function that represents a scalar binary
operator into one that computes the \code{\link{kronecker}}
version, but with dimensions named in a way that is more
convenient for use with \code{macpan2}.
}
