% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_data.R
\name{mp_ledgers}
\alias{mp_ledgers}
\title{#' @export
print.LedgerData = function(x, ...) {
print(x$frame, row.names = FALSE)
}}
\usage{
mp_ledgers(...)
}
\description{
#' @export
print.LedgerData = function(x, ...) {
print(x$frame, row.names = FALSE)
}
}
