% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{mp_log_normal}
\alias{mp_log_normal}
\title{Log-Normal Distribution}
\usage{
mp_log_normal(location = NULL, sd = 1)
}
\arguments{
\item{location}{Location parameter for the log-normal distribution.}

\item{sd}{Standard deviation parameter for the normal distribution. The
standard deviation of the log-normal distribution is \code{log_sd = log(sd)}.}
}
\description{
Log-Normal Distribution
}
