% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{mp_neg_bin}
\alias{mp_neg_bin}
\title{Negative Binomial Distribution}
\usage{
mp_neg_bin(location = NULL, disp)
}
\arguments{
\item{disp}{Dispersion parameter.}
}
\description{
Negative Binomial Distribution
}
