% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{mp_normal}
\alias{mp_normal}
\title{Normal Distribution}
\usage{
mp_normal(location = DistrParam("location"), sd)
}
\arguments{
\item{location}{Location parameter. Only necessary if used as a prior
distribution. If it is used as a likelihood component the location
parameter will be taken as the simulated variable being fitted to data,
and so this \code{location} parameter should be left to the default.}

\item{sd}{Standard deviation parameter.}
}
\value{
DistrSpec
}
\description{
Normal Distribution
}
