% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_list_generators.R
\name{mp_per_capita_flow}
\alias{mp_per_capita_flow}
\title{Per-Capita Flow (Experimental)}
\usage{
mp_per_capita_flow(from, to, rate)
}
\arguments{
\item{from}{String giving the name of the compartment from which the flow
originates.}

\item{to}{String giving the name of the compartment to which the flow is
going.}

\item{rate}{A two-sided formula with the left-hand-side giving the name of
the absolute flow rate per unit time-step and the right-hand-side giving
an expression for the per-capita rate of flow from \code{from} to \code{to}.}
}
\description{
Per-Capita Flow (Experimental)
}
