% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_list_generators.R
\name{mp_per_capita_flow}
\alias{mp_per_capita_flow}
\alias{mp_per_capita_inflow}
\title{Per-Capita Flow (Experimental)}
\usage{
mp_per_capita_flow(from, to, rate, abs_rate = NULL)

mp_per_capita_inflow(from, to, rate, abs_rate = NULL)
}
\arguments{
\item{from}{String giving the name of the compartment from which the flow
originates.}

\item{to}{String giving the name of the compartment to which the flow is
going.}

\item{rate}{String giving the expression for the per-capita flow rate.
Alternatively, for back compatibility, a two-sided formula with the
left-hand-side giving the name of the absolute flow rate per unit time-step
and the right-hand-side giving an expression for the per-capita rate of
flow from \code{from} to \code{to}.}

\item{abs_rate}{String giving the name for the absolute flow rate,
which will be computed as \code{from * rate}. If a formula is passed to
\code{rate} (not recommended), then this \code{abs_rate_name} argument will be ignored.}
}
\description{
Per-Capita Flow (Experimental)
}
\section{Functions}{
\itemize{
\item \code{mp_per_capita_inflow()}: Only flow into the \code{to} compartment, and
do not flow out of the \code{from} compartment.

}}
