% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrator_arg_constructors.R
\name{mp_rbf}
\alias{mp_rbf}
\title{Fit a Time-Varying Parameter with Radial Basis Functions}
\usage{
mp_rbf(tv, dimension, initial_weights, seed = 1L, prior_sd = 1)
}
\arguments{
\item{tv}{String giving the name of the parameter.}

\item{dimension}{Number of bases.}

\item{initial_weights}{Optional vector with \code{dimensions} elements. These
are the parameters that are fitted and determine how \code{tv} varies with
time.}

\item{seed}{Optional random seed to use to generate the \code{initial_weights}
if they are not provided.}

\item{prior_sd}{Optional prior standard deviation default value for radial
basis function coefficients, defaults to 1.}
}
\description{
Pass the output of this function to the \code{tv} argument of
\code{\link{mp_tmb_calibrator}} to model time variation of
a parameter with flexible radial basis functions.
}
