% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_tmb_model_spec.R
\name{mp_read_rds}
\alias{mp_read_rds}
\title{Read Serialized Model Specification}
\usage{
mp_read_rds(filename)
}
\arguments{
\item{filename}{Path to a saved model specification object.}
}
\description{
Uses \code{\link{readRDS}} to read in a saved model specification
created using a function like \code{\link{mp_tmb_model_spec}}, and
updates this specification using \code{\link{mp_version_update}} so that
it is compatible with the installed version of \code{macpan2}. To save
a model specification, just use the base \code{R} function \code{\link{saveRDS}}.
}
