% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_cal_time.R
\name{mp_sim_bounds}
\alias{mp_sim_bounds}
\title{Simulation Bounds (Experimental)}
\usage{
mp_sim_bounds(sim_start, sim_end, time_scale)
}
\arguments{
\item{sim_start}{Start time of each simulation.}

\item{sim_end}{End time of each simulation.}

\item{time_scale}{Qualitative description of the size of a time step.
currently only \code{"steps"}, \code{"daily"}, and \code{"weekly"} are allowed,
and but \code{"steps"} is the only recommended version as the other two
are poorly tested and will throw a warning. The recommended \code{"steps"}
option assumes that positive integers are used to indicate a particular
point in the simulation.}
}
\description{
Set the simulation bounds (start time and end time) for a
calibration. This is used to override the default simulation
bounds taken from the observed data passed to
\code{\link{mp_tmb_calibrator}}. the first date is when the first
simulated time step (chosen to be before the first data point
so that infectious individuals can be built up) and
the second date is the last simulated time step (chosen
to be after the last data point so that there can be
a forecast period). the last argument gives the
scale of a single time step (in this case it should
always be daily).
}
