% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp.R
\name{mp_subset}
\alias{mp_subset}
\alias{mp_setdiff}
\title{Subset of Indexes}
\usage{
mp_subset(x, ...)

mp_setdiff(x, ...)
}
\arguments{
\item{x}{Model index.}

\item{...}{Name-value pairs. The names are columns (or sets of columns
using dot-concatenation) in \code{x} and the values are character vectors
that refer to labels with respect to those columns. These values
determine the resulting subset.}
}
\description{
Take a subset of the rows of an index table (see \code{\link{mp_index}})
to produce another index table. The \code{mp_subset} function gives rows that
match a certain criterion and \code{mp_setdiff} gives rows that do not match.
}
\seealso{
Other functions that return index tables
\code{\link{mp_cartesian}()},
\code{\link{mp_index}()},
\code{\link{mp_rename}()},
\code{\link{mp_union}()}
}
\concept{indexes}
