% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_model_editors.R
\name{mp_tmb_insert_backtrans}
\alias{mp_tmb_insert_backtrans}
\alias{mp_tmb_implicit_backtrans}
\title{Insert Back Transformations of Model Parameters}
\usage{
mp_tmb_insert_backtrans(
  model,
  variables = character(),
  transformation = mp_log
)

mp_tmb_implicit_backtrans(model, variables = character())
}
\arguments{
\item{model}{TMB model spec object produced using
\code{\link{mp_tmb_library}} or \code{\link{mp_tmb_model_spec}}.}

\item{variables}{Character vector of parameters to back transform.}

\item{transformation}{A transformation object such as \code{\link{mp_log}},
which is the default. See the help page for \code{\link{mp_log}} for
available options.}
}
\value{
A new model specification object with expressions for the
untransformed (or back transformed) parameters at the beginning of the
\code{before} phase. The transformed version of the parameter is also
added to the defaults and are identified with a prefixed name (e.g.,
\code{log_beta} if \code{beta} is log transformed).
}
\description{
Insert Back Transformations of Model Parameters
}
\section{Functions}{
\itemize{
\item \code{mp_tmb_implicit_backtrans()}: Insert parameter transformations
implicitly by pre-pending the name of the transformations in front of the
names of the \code{variables} that you ask for (e.g., \code{"log_case_reports"}) even
if this variable is not in the model as long as the base name (e.g.,
\code{"case_reports"}) is.

}}
\examples{
init_si = ("starter_models"
  |> mp_tmb_library("si", package = "macpan2")
  |> mp_tmb_insert_backtrans("beta", mp_log)
  |> mp_initial_list()
)
print(init_si$log_beta)
print(log(init_si$beta))

}
\seealso{
\code{\link[=mp_tmb_insert_trans]{mp_tmb_insert_trans()}}
}
