% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_model_editors.R
\name{mp_tmb_insert_reports}
\alias{mp_tmb_insert_reports}
\title{Insert Reports}
\usage{
mp_tmb_insert_reports(
  model,
  incidence_name,
  report_prob,
  mean_delay,
  cv_delay,
  reports_name = sprintf("reported_\%s", incidence_name),
  report_prob_name = sprintf("\%s_report_prob", incidence_name),
  mean_delay_name = sprintf("\%s_mean_delay", incidence_name),
  cv_delay_name = sprintf("\%s_cv_delay", incidence_name)
)
}
\arguments{
\item{incidence_name}{Name of the incidence variable to be transformed.}

\item{report_prob}{Value to use for the reporting probability; the
proportion of cases that get reported.}

\item{mean_delay}{Mean of the Gamma distribution of reporting delay times.}

\item{cv_delay}{Coefficient of variation of the Gamma distribution of
reporting delay times.}

\item{reports_name}{Name of the new reports variable.}

\item{report_prob_name}{Name of the variable containing \code{report_prob}.}

\item{mean_delay_name}{Name of the variable containing \code{mean_delay}.}

\item{cv_delay_name}{Name of the variable containing \code{cv_delay}.}
}
\description{
A version of \code{\link{mp_tmb_insert}} making it more convenient to
transform an incidence variable into a reports variable, which accounts
for reporting delays and under-reporting. This new reports variable is
a convolution of the simulation history of an incidence variable with
a kernel that is proportional to a Gamma distribution of reporting
delay times.
}
