% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_library.R
\name{mp_tmb_library}
\alias{mp_tmb_library}
\title{TMB Library}
\usage{
mp_tmb_library(..., package = NULL)
}
\arguments{
\item{...}{File path components pointing to a directory that
contains an R script that creates an object called \code{spec}, which
is produced by \code{\link{mp_tmb_model_spec}}.}

\item{package}{If \code{NULL}, \code{\link{file.path}} is used
to put together the \code{...} components but if \code{package}
is the name of a package (as a character string) then
\code{\link{system.file}} is used to put together the \code{...}
components.}
}
\description{
Get a TMB model specification from a model library.
}
\examples{
mp_tmb_library(
    "starter_models"
  , "si"
  , package = "macpan2"
)

}
\seealso{
\code{\link[=show_models]{show_models()}}
}
