% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_library.R
\name{mp_tmb_library}
\alias{mp_tmb_library}
\alias{mp_tmb_entire_library}
\title{Read Item from a Model Library}
\usage{
mp_tmb_library(..., package = NULL, alternative_specs = FALSE)

mp_tmb_entire_library()
}
\arguments{
\item{...}{File path components pointing to a directory that
contains an R script that creates an object called \code{spec}, which
is produced by \code{\link{mp_tmb_model_spec}}.}

\item{package}{If \code{NULL}, \code{\link{file.path}} is used
to put together the \code{...} components but if \code{package}
is the name of a package (as a character string) then
\code{\link{system.file}} is used to put together the \code{...}
components.}

\item{alternative_specs}{If \code{TRUE}, return a list of alternative
specification objects. For models without alternatives this will cause
the return value to be a list with one element containing a spec object.}
}
\description{
Get a TMB model specification from a model library.
}
\section{Functions}{
\itemize{
\item \code{mp_tmb_entire_library()}: List of one model specification for each model
in the library.

}}
\examples{
mp_tmb_library(
    "starter_models"
  , "si"
  , package = "macpan2"
)

}
\seealso{
\code{\link[=mp_show_models]{mp_show_models()}}
}
\concept{create-model-spec}
