% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_to_tmb.R
\name{mp_tmb_simulator}
\alias{mp_tmb_simulator}
\title{TMB Simulator from Dynamic Model}
\usage{
mp_tmb_simulator(...)
}
\arguments{
\item{dynamic_model}{Object product by \code{\link{dynamic_model}}.}

\item{vectors}{Named list of named vectors as initial values for the
simulations that are referenced in the expression list in the dynamic model.}

\item{unstruc_mats}{= Named list of objects that can be coerced to
numerical matrices that are used in the expression list of the
dynamic model.}
}
\description{
TMB Simulator from Dynamic Model
}
