% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_methods.R
\name{mp_tmbstan_coef}
\alias{mp_tmbstan_coef}
\title{Model Coefficient Table with stan}
\usage{
mp_tmbstan_coef(model, tmbstan_args = list(), ...)
}
\arguments{
\item{model}{Object that contains information about model coefficients.}

\item{...}{Arguments to pass onto the \code{broom.mixed::tidy.stanfit} method.}
}
\description{
Leverages the \code{tmbstan} and \code{broom.mixed} packages to generate
MCMC-based coefficient tables.
}
