% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_model.R
\name{mp_trajectory}
\alias{mp_trajectory}
\title{Trajectory}
\usage{
mp_trajectory(model)
}
\arguments{
\item{model}{A dynamical model simulator produced by
\code{\link{mp_simulator}}.}
}
\description{
Return simulations of the trajectory of the output
variables of a dynamical model simulator.
}
