% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_frame.R
\name{mp_vars}
\alias{mp_vars}
\alias{mp_state_vars}
\alias{mp_flow_vars}
\alias{mp_state_flow_vars}
\title{State and Flow Variable Names}
\usage{
mp_state_vars(spec, topological_sort = FALSE, loops = "^$", prefix = "")

mp_flow_vars(spec, topological_sort = FALSE, loops = "^$", trans = "")

mp_state_flow_vars(spec, topological_sort = FALSE, loops = "^$", trans = "")
}
\arguments{
\item{spec}{Model specification (\code{\link{mp_tmb_model_spec}}).}

\item{topological_sort}{Should the states and flows be
\href{https://en.wikipedia.org/wiki/Topological_sorting}{topologically sorted} to
respect the main direction of flow? The default is no topological sorting,
which differs from \code{\link{mp_flow_frame}}.}

\item{loops}{Pattern for matching the names of flows that make
the flow model not a DAG, which is a critical assumption when topologically
sorting the order of states and flows in the output. This is only relevant if
\code{topological_sort} is used.}

\item{trans}{Add a prefix to the names for indicating if a transformed
version of the variables is preferred.}
}
\value{
Character vector of names of all state and/or flow variables that
have been explicitly represented in the model using functions like
\code{\link{mp_per_capita_flow}}.
}
\description{
Get the state and/or flow variables in a model specification.
}
\section{Functions}{
\itemize{
\item \code{mp_state_vars()}: Return character vector of all state variables.

\item \code{mp_flow_vars()}: Return the names of all variables that contain
the absolute flow between compartments.
The absolute flow is the magnitude of a flow per time step.

\item \code{mp_state_flow_vars()}: Union of \code{mp_state_vars()} and \code{mp_flow_vars()}.

}}
\examples{
si = mp_tmb_library("starter_models", "si", package = "macpan2")
(si
  |> mp_simulator(time_steps = 5L, mp_state_vars(si))
  |> mp_trajectory()
)

}
