% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_tmb_model_spec.R
\name{mp_version_update}
\alias{mp_version_update}
\title{Version Update}
\usage{
mp_version_update(spec)
}
\arguments{
\item{spec}{Object produced by \code{\link{mp_tmb_model_spec}} or another
function that produces the same type of object.}
}
\description{
Update the \code{macpan2} version associated with a model specification.
}
