% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_expr.R
\name{parse_expr_list}
\alias{parse_expr_list}
\title{Parse Expression List}
\usage{
parse_expr_list(
  expr_list,
  valid_vars,
  valid_literals = numeric(0L),
  offset = 0L
)
}
\arguments{
\item{expr_list}{List of one-sided formulas.}

\item{valid_vars}{Named list of numerical matrices that can
be referred to in the formulas.}
}
\description{
Parse a list of one-sided formulas representing expressions
in a compartmental model.
}
