% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_partitions.R
\name{Partition}
\alias{Partition}
\title{Partition}
\usage{
Partition(frame)
}
\arguments{
\item{frame}{Data frame representing the partition.}
}
\description{
Create object for manipulating partitions, which are sets of
labels for representing and naming model entities.
}
