% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{Reader}
\alias{Reader}
\alias{CSVReader}
\alias{JSONReader}
\alias{TXTReader}
\alias{RReader}
\alias{NULLReader}
\title{Reader}
\usage{
Reader(...)

CSVReader(...)

JSONReader(...)

TXTReader(...)

RReader(...)

NULLReader(...)
}
\arguments{
\item{...}{Character vectors giving path components to the file to be read.}
}
\description{
Construct objects for reading data.
}
\section{Functions}{
\itemize{
\item \code{CSVReader()}: Read CSV files.

\item \code{JSONReader()}: Read JSON files.

\item \code{TXTReader()}: Read TXT files.

\item \code{RReader()}: Read R files.

\item \code{NULLReader()}: Placeholder reader that always returns \code{NULL}.

}}
