% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_sims.R
\name{simple_sims}
\alias{simple_sims}
\title{Simple Iterated Simulation}
\usage{
simple_sims(iteration_exprs, time_steps, int_vecs = list(), mats = list())
}
\arguments{
\item{iteration_exprs}{List of expressions to pass to the
\href{https://canmod.github.io/macpan2/articles/cpp_side}{engine}. The expressions
are only allowed to use valid \code{\link{engine_functions}}.}

\item{time_steps}{Number of time steps to iterate.}

\item{...}{Named list \code{\link{numeric}} objects that can be coerced to
\code{\link{numeric}} \code{\link{matrix}} objects. These matrices can be
referred to by name in the expressions in \code{iteration_exprs}.}
}
\value{
A data frame with the simulation results returned by the \verb{$report()}
method in \code{\link{TMBSimulator}}.
}
\description{
Simple Iterated Simulation
}
