% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_model_constructors.R
\name{SimulatorConstructor}
\alias{SimulatorConstructor}
\title{Simulator Constructor}
\usage{
SimulatorConstructor(model_directory, ...)
}
\arguments{
\item{model_directory}{A string giving a path to a directory containing model definition files}

\item{...}{Arguments to pass to \link{\code{TMBModel}}}
}
\description{
Simulator Constructor
}
