% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_to_models.R
\name{StandardExpr}
\alias{StandardExpr}
\title{Standard Expressions}
\usage{
StandardExpr(model)
}
\arguments{
\item{model}{Object created by \code{\link{Model}}.}
}
\description{
Evaluate standard model expressions.
}
