% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled_partitions.R
\name{to_names}
\alias{to_names}
\title{To Names}
\usage{
to_names(x)
}
\arguments{
\item{x}{Object to convert to labels.}
}
\value{
Character vector that can be used as names.
}
\description{
Convert objects to names, which are vectors that cannot be dotted.
}
