% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_trans.R
\name{Transform}
\alias{Transform}
\alias{Log}
\alias{Logit}
\title{Transform}
\usage{
Transform(variable, default = NULL, prefix = "")

Log(variable, default = NULL)

Logit(variable, default = NULL)
}
\arguments{
\item{variable}{Character string giving a variable in the model.}

\item{default}{Default value for the untransformed variable. If \code{NULL}
(the default) this value is taken from the initial value in the model
containing the transformation.}

\item{prefix}{Character string to use to distinguish the transformed
version of the \code{variable}.}
}
\description{
Transform
}
\section{Functions}{
\itemize{
\item \code{Log()}: Log transformation

\item \code{Logit()}: Logit transformation

}}
