% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_trans.R
\name{Transform}
\alias{Transform}
\alias{Log}
\alias{Logit}
\title{Transform}
\usage{
Transform(variable, prefix)

Log(variable)

Logit(variable)
}
\arguments{
\item{variable}{Character string giving a variable in the model.}

\item{prefix}{Character string to use to distinguish the transformed
version of the \code{variable}.}
}
\description{
Transform
}
\section{Functions}{
\itemize{
\item \code{Log()}: Log transformation

\item \code{Logit()}: Logit transformation

}}
